% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_NMD.R
\name{abbrevWords}
\alias{abbrevWords}
\title{Abbreviates StoX project names.}
\usage{
abbrevWords(x, p = 1/2, collapse = "_", keep = c("capital", "numeric",
  "punct"), sub = 0, abbrev = TRUE)
}
\arguments{
\item{x}{A vector of strings.}

\item{p}{The exponent in the abbreviation function. The number of characters remaining (n) after possibly capital letters and numbers have been fixed in the abbreviated string, are reduced in number according to round(n^p).}

\item{collapse}{The character to separate words by.}

\item{keep}{Keywords for what to keep in the abbreviated strings. Including "punct" ensures that punctuation characters are left untouched, but still considered as separators between words.}

\item{sub}{A value to add to the number of characters in a single abbreviated word. Set this to say -1 to remove wery whort words such as "in".}
}
\description{
Abbreviates StoX project names.
}
\examples{
cs <- getNMDinfo("cs", recursive = FALSE)
sts <- getNMDinfo("sts", recursive = FALSE)
abbr <- list(cbind(cs, abbrevWords(cs)), cbind(sts, abbrevWords(sts)))
abbr
lapply(abbr, nchar)

}
\keyword{internal}
