% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_base.r
\name{abbrMatch}
\alias{abbrMatch}
\title{Get subsequently plotting variables used in plotting functions in Rstox (using abbreviated string mathes obtained by abbrMatch())}
\usage{
abbrMatch(x, table, ignore.case = FALSE)
}
\arguments{
\item{x}{A single string naming a variable.}

\item{table}{A vector of stings to match \code{x} against.}

\item{ignore.case}{Logical: if TRUE, ignore case when matching.}
}
\description{
These functions are used in the plotting and reporting functions getPlots() and getReports() and dependent functions.
}
\keyword{internal}

