% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logbooks.R
\name{adjustGearLandingsWithLogbooksReca}
\alias{adjustGearLandingsWithLogbooksReca}
\title{Adjusts landings for Reca}
\usage{
adjustGearLandingsWithLogbooksReca(
  landingsStox,
  logbooksLst,
  gearTable,
  gearCode,
  minVesselSize = 15
)
}
\arguments{
\item{gearTable}{table of gear mappings. As exported to processData by the ECA template stox-baseline.}

\item{gearCode}{code identifying the gear (as in gearTabe$Covariate).}

\item{minVesselSize}{the minimal vessel size that logbook corrections should be applied to}

\item{landingStox}{the landings, formatted as in stoxExport$landing, stored by \code{\link[Rstox]{prepareRECA}}}

\item{logbookLst}{Aggregated logbook on lst format, as parsed by \code{\link[Rstox]{readLstFile}}}
}
\value{
landings, formatted as 'landingStox'
}
\description{
Adjust landings for selected gears.
}
\details{
Note: This is prototype functionaltiy consider for proper inclusion in later versions of StoX
 Note: Ideally such adjustments are done on aggregated formats, but are included at an earlier stage here because of existing software structures
 Note: Some landings may be introduced by resampling, and identifying information may be incorrect (e.g. sale notes ids)
 
 Calculates proportions of catch of each species, in quarters and main areas (hovedområde) for a specified gear
 in logbooks, and adjust the landed weights accordingly.
 
 Adjustments are done by uniformly scaling all landings within a cell.
 
 For instance trawl-landings, or other landing from multi-day trips can be adjusted to
 reflect the higher spatial and temporal resolution in logbooks as opposed to landings (sales notes)
 
 Fits the format for landings used internally in the Reca scripts.
}
