% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logbooks.R
\name{adjustLandings}
\alias{adjustLandings}
\title{Adjust landings}
\usage{
adjustLandings(landings, proportions, totalcell, subcell, weight)
}
\arguments{
\item{landings}{landings to be adjusted}

\item{proportions}{data.table of proportions in cells. Proportion should be in a column named 'fraction'}

\item{subcell}{vector of strings, identfies cells (in 'landings' and 'proportions'), relative to totalcells which ar to be adjusted}

\item{weight}{string identifying the column where landed weight is given in 'landings'}

\item{totalcells}{vector of strings, identifies cells (in 'landings' and 'proportions') within which total catch is to be kept constant.}
}
\value{
adjusted landings, formatted as 'landings'
}
\description{
Adjust weights of landings according to pre-calculated proportions
}
\details{
Note: This is prototype functionaltiy consider for proper inclusion in later versions of StoX
 Note: Ideally such adjustments are done on aggregated formats, but are included at an earlier stage here because of existing software structures
 Note: Some landings may be introduced by resampling, and identifying information may be incorrect (e.g. sale notes ids)
 
 A cell is a combination of categorical values, e.g. quarter and area.
 'totalcell' are cells whose totals should be unchanged by by the adjustment.
 'subcells' are cells which should be assigned a proportion of the in each totalcell. 
 These are defined relative to totalcell.
 
 Any cells with with totalcell variables in 'landings' that are not in logbooks are left untouched
 Any totalcell in proportions, but not in landings will raise an error.
 Any subcells in 'proportions', but not in 'landings' will be added by sampling one landing from the same totalcell, and assign the appropriate fraction of the total catch
 Any subcells in 'landings', but not in 'proportions will treated as if the proportions for these cells are zero.
}
