% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logbooks.R
\name{adjustRecaSpatialTemporal}
\alias{adjustRecaSpatialTemporal}
\title{Adjust landings with logbooks}
\usage{
adjustRecaSpatialTemporal(
  landingsStox,
  logbook,
  processDataGear,
  processDataTemporal,
  processDataSpatial,
  gearSelection,
  temporalCovariate = "temporal",
  spatialCovariate = "spatial",
  gearCovariate = "gearfactor",
  minVesselSize = 15,
  logbookActivityCodes = c("I fiske")
)
}
\arguments{
\item{landingsStox}{landings as passed between the Reca-scripts. E.g. as saved by \code{\link[Rstox]{prepareRECA}}}

\item{logbook}{logbooks as parser by \code{\link[Rstox]{readErsFile}}}

\item{processDataGear}{processdata gerfactor as exported from Stox-baseline}

\item{processDataTemporal}{processdata temporal}

\item{processDataSpatial}{processdata stratumpolygon as exported from Stox-baseline}

\item{gearSelection}{identifies the gear to clean for. One of the options on processDataGear$Covariate}

\item{temporalCovariate}{the name of the temporal covariate}

\item{spatialCovariate}{the name of the spatial covariate}

\item{gearCovariate}{the name of the gear covariate}

\item{minVesselSize}{minimal size in meters for vessels to include in logbook adjusted data. Vessels without length will be assumed smaller than this threshold}

\item{logbookActivityCodes}{vector specifying the activity codes to include from logbooks (field AKTIVITET).}
}
\value{
landings, formatted as landingStox
}
\description{
Adjusts total weight in spatial and temporal covariates for a given gear, based on logbook data.
}
\details{
Note: This is prototype functionaltiy consider for proper inclusion in later versions of StoX
 Note: Ideally such adjustments are done on aggregated formats, but are included at an earlier stage here because of existing software structures
 Note: Some landings may be introduced by resampling
 Note: Some columns on resamled records may be missing or incorrect, 
 Note: but the covariates, 'sistefangstdato', 'artkode', and weight parameters are set in representative fashion.
 
 Logbook-adjustment ensures that:
 \itemze{
  \item total weight in adjusted landings are the same as total weight before adjustment
  \item the relative proportions of the part of the landings that are covered by logbooks are the same as in logbooks
 }
 The portion covered by logbooks are considered identified by 'gearselection' and 'minVesselSize'
}
