% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_reports.r
\name{aggregateBySpeciesCategory}
\alias{aggregateBySpeciesCategory}
\title{Generate species matrix with biotic stations as rows and station information and the species present in the \code{ref} file as columns. One matrix generated per variable \code{var}.}
\usage{
aggregateBySpeciesCategory(projectName, ref, specVar = "noname",
  specVar.bio = specVar, specVar.ref = specVar, catVar = "SpecCat",
  bioticProc = "FilterBiotic", stationVar = c("cruise", "serialno"),
  var = c("weight", "count"), na.as = 0, drop.out = TRUE, close = TRUE,
  msg = FALSE, ...)
}
\arguments{
\item{projectName}{The name or full path of the project, a baseline object (as returned from \code{\link{getBaseline}} or \code{\link{runBaseline}}, og a project object (as returned from \code{\link{openProject}}).}

\item{ref}{The path to a file linking species to species categories, specifically named by \code{specVar} or \code{specVar.ref} and \code{catVar}. The file must be a CSV (comma separated) file with UTF-8 encoding.}

\item{specVar, specVar.bio, specVar.ref}{Names of the species columns in the data and in the \code{ref} file. When only specifying \code{specVar}, a common column name is assumed.}

\item{catVar}{The name of the species category column in the \code{ref} file.}

\item{bioticProc}{The process from which the biotic data should be extracted from the project.}

\item{stationVar}{The names of the columns of the biotic data identifying the biotic stations. These columns will be concatenated.}

\item{var}{The names of the columns for which the species matrices should be generated.}

\item{na.as}{The value to use for missing data (e.g., species categories that are not present in a station).}

\item{drop.out}{Logical: If TRUE (default) unlist if only one variable is given in \code{var}.}

\item{close}{Logical: If TRUE (default) close the project after reading the data.}

\item{...}{Parameters passed on to getBaseline(), e.g. for changing the baseline parameters (adding filters).}
}
\value{
A list of matrices (dropped to a matrix if only one variable is specified in \code{var}) with stations as rows and station information and the species present in the \code{ref} file as columns.
}
\description{
Generate species matrix with biotic stations as rows and station information and the species present in the \code{ref} file as columns. One matrix generated per variable \code{var}.
}
\keyword{internal}
