% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_base.r
\name{as.matrix_full}
\alias{as.matrix_full}
\alias{as.dataFrame_full}
\title{Convert list to matrix or data.frame and generate missing values}
\usage{
as.matrix_full(x, stringsAsFactors = FALSE)

as.dataFrame_full(x, stringsAsFactors = FALSE)
}
\arguments{
\item{x}{A list of vectors of one dimension with column names or names.}
}
\description{
Convert a list of vectors with variable length to a matrix with all possible variables in the columns, and with missing values
}
\keyword{internal}
