% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_eca.r
\name{baseline2ECA}
\alias{baseline2ECA}
\alias{baseline2eca}
\title{Prepare data from a project to the ECA model}
\usage{
baseline2ECA(projectName, biotic = "BioticCovData",
  landing = "LandingCovData", temporal = NULL, gearfactor = NULL,
  spatial = NULL)

baseline2eca(projectName, biotic = "BioticCovData",
  landing = "LandingCovData", temporal = NULL, gearfactor = NULL,
  spatial = NULL)
}
\arguments{
\item{projectName}{The name or full path of the project, a baseline object (as returned from getBaseline() or runBaseline()), or a project object (as returned from open).}

\item{biotic}{The process from which the biotic data are extracted, conventionally the BioticCovData process.}

\item{landing}{The process from which the landing data are extracted, conventionally the LandingCovData process.}

\item{temporal}{Optional definition of the temporal covariate (not yet implemented).}

\item{gearfactor}{Optional definition of the gearfactor covariate (not yet implemented).}

\item{spatial}{Optional definition of the spatial covariate (not yet implemented).}
}
\value{
A reference to the StoX Java baseline object
}
\description{
This function reads data from a baseline run and converts to a list of data used by the ECA model in the Reca package.
}
\examples{
# Create the test project:
createProject("Test_Rstox", files=system.file("extdata", "Test_Rstox", package="Rstox"), ow=TRUE)
system.time(eca <- baseline2ECA("Test_Rstox"))
# Show the covariate definitions:
eca$resources
# Show the eca object:
str(eca)

}

