% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_boostrap.r
\name{bootstrapBioticAcoustic}
\alias{bootstrapBioticAcoustic}
\title{(Internal) Bootstrap biotic stations}
\usage{
bootstrapBioticAcoustic(baseline, assignments, psuNASC, stratumNASC,
  resampledNASC, parameters, ...)
}
\arguments{
\item{baseline}{StoX Java baseline object}

\item{assignments}{Trawl assignment from baseline}

\item{psuNASC}{MeanNASC from baseline}

\item{stratumNASC}{Strata NASC estimates from getNASCDistr(baseline)}

\item{resampledNASC}{Resampled NASC distribution}

\item{parameters}{Parameters set by user in Stox; 
\describe{
\item{parameters$nboot}{Number of bootstrap replicates}
\item{parameters$seed}{The seed for the random number generator (used for reproducibility)}
}}

\item{...}{Used for backwards compatibility.}
}
\value{
list with n = nboot data.frames containing results from baseline and bootstrap replicates
}
\description{
Resample (bootstrap) trawl stations based on survey (Cruise) number and station numbers (SerialNo) to estimate uncertainty in estimates.
}

