% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_boostrap.r
\name{bootstrapBioticSweptArea}
\alias{bootstrapBioticSweptArea}
\title{(Internal) Bootstrap biotic stations Swept Area version}
\usage{
bootstrapBioticSweptArea(baseline, assignments, parameters, ...)
}
\arguments{
\item{baseline}{StoX Java baseline object.}

\item{assignments}{Trawl assignment from baseline.}

\item{parameters}{R-object with parameters read from rprocess.txt.
\describe{
\item{parameters$nboot}{Number of bootstrap replicates}
\item{parameters$seed}{The seed for the random number generator (used for reproducibility)}
}}

\item{...}{Used for backwards compatibility.}
}
\value{
list with n = nboot data.frames containing results from baseline and bootstrap replicates
}
\description{
Resample (bootstrap) trawl stations based on survey (Cruise) number and station numbers (SerialNo) to estimate uncertainty in estimates
}

