% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_boostrap.r
\name{bootstrapParallel}
\alias{bootstrapParallel}
\title{(Internal) Bootstrap biotic stations and acoustic data}
\usage{
bootstrapParallel(projectName, assignments, psuNASC = NULL,
  stratumNASC = NULL, resampledNASC = NULL, nboot = 5,
  startProcess = "TotalLengthDist", endProcess = "SuperIndAbundance",
  outProcess = "SuperIndAbundance", seed = 1, cores = 1,
  baseline = NULL, msg = TRUE, parameters = list())
}
\arguments{
\item{projectName}{The name or full path of the project, a baseline object (as returned from getBaseline() or runBaseline()), og a project object (as returned from open).}

\item{assignments}{Trawl assignment from baseline.}

\item{psuNASC}{MeanNASC from baseline.}

\item{stratumNASC}{Strata NASC estimates from getNASCDistr(baseline).}

\item{resampledNASC}{Resampled NASC distribution.}

\item{nboot}{Number of bootstrap replicates.}

\item{seed}{The seed for the random number generator (used for reproducibility).}

\item{cores}{An integer giving the number of cores to run the bootstrapping over.}

\item{baseline}{Optional: a StoX baseline object returned from runBaseline().}

\item{msg}{Logical: if TRUE print messages from runBaseline().}

\item{parameters}{Parameters set by user in Stox (only kept for compatibility with older versions);.
\describe{
\item{parameters$nboot}{Number of bootstrap replicates}
\item{parameters$seed}{The seed for the random number generator (used for reproducibility)}
}}
}
\value{
list with (1) the abundance by length in the orginal model, (2) the abundance by length in the bootstrap run, (3) the abundance by super individuals in the orginal model, (4) the abundance by super individuals in the bootstrap run
}
\description{
Resample (bootstrap) trawl stations based on survey (Cruise) number and station numbers (SerialNo) to estimate uncertainty in estimates.
}
\keyword{internal}
