% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logbooks.R
\name{calculateCatchProportions}
\alias{calculateCatchProportions}
\title{Calculates catch proportions}
\usage{
calculateCatchProportions(
  logbook,
  totalcell = c("REGM", "FANGSTART"),
  subcell = c("LOKASJON_START"),
  weight = "RUNDVEKT",
  na.rm = F
)
}
\arguments{
\item{logbook}{data.table with logbook records, as parsed by \code{\link[Rstox]{readErsFile}}}

\item{totalcell}{vector of strings identifying column names in 'logbook', identifying cells with totals}

\item{subcell}{vector of strings identifying column names in 'logbook', identifying cells which are to be assigned a proportion}

\item{weight}{string identifying the column containing weights.}
}
\value{

}
\description{
Calculates the proportion of catch that was caught in specified cells,
 according to logbook (ERS) records.
}
\details{
A cell is a combination of categorical values, e.g. quarter and area.
 'totalcell' are cells which should have total fraction of 1.
 'subcells' are cells which should be assigned a proportion of the total. 
 These are defined relative to totalcell.
 
 E.g: In order to get the fraction of a certain species, caught at each location for a vessel in a year:
 calculateCatchProportions(logbook, totalcell=c("REGM", "FANGSTART"), subcell=c("LOKASJON_START")),
 specifying that the fractions sum to one over REGM (vessel) and FANGSTART (species), 
 and that each LOKASJON_START (location) should be assigned the fraction of the catch caught there.
}
