% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RECA_format_checks.R
\name{check_columns_present}
\alias{check_columns_present}
\title{checks that column names are present on datamatrix}
\usage{
check_columns_present(datamatrix, columns)
}
\description{
checks that column names are present on datamatrix
}
\keyword{internal}
