% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RECA_data_checks.R
\name{check_common_covariates_for_NAs}
\alias{check_common_covariates_for_NAs}
\title{Checks for data issues wrp to running RECA}
\usage{
check_common_covariates_for_NAs(stoxexport)
}
\description{
Checks for NAs in columns commonly used for setting covariates
}
\keyword{internal}
