% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RECA_format_checks.R
\name{check_covariates}
\alias{check_covariates}
\title{checks that specification of covariates are OK}
\usage{
check_covariates(modelobject)
}
\description{
checks that specification of covariates are OK
}
\keyword{internal}
