% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RECA_format_checks.R
\name{check_landings_cov}
\alias{check_landings_cov}
\title{checks formatting on landing cov-matrices}
\usage{
check_landings_cov(cov)
}
\description{
checks formatting on landing cov-matrices
}
\keyword{internal}
