% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RECA_format_checks.R
\name{check_none_missing}
\alias{check_none_missing}
\title{columns does not have missing values}
\usage{
check_none_missing(datamatrix, columns)
}
\description{
columns does not have missing values
}
\keyword{internal}
