% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RECA_format_checks.R
\name{checkAgeLength}
\alias{checkAgeLength}
\title{checks that agelenght is configured correctly}
\usage{
checkAgeLength(agelength, num_tolerance = 1e-10)
}
\description{
checks that agelenght is configured correctly
}
\keyword{internal}
