% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RECA_format_checks.R
\name{checkCovariateConsistency}
\alias{checkCovariateConsistency}
\title{checks that covariates are compatible between model and landings}
\usage{
checkCovariateConsistency(modelobj, landingscov)
}
\description{
checks that covariates are compatible between model and landings
}
\keyword{internal}
