% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RECA_format_checks.R
\name{checkLandings}
\alias{checkLandings}
\title{checks that landings are specified correctly}
\usage{
checkLandings(landings)
}
\description{
checks that landings are specified correctly
}
\keyword{internal}
