% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RECA_format_checks.R
\name{checkWeightLength}
\alias{checkWeightLength}
\title{checks that weightlenght is configured correctly}
\usage{
checkWeightLength(weightlength, landings)
}
\description{
checks that weightlenght is configured correctly
}
\keyword{internal}
