% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_xml.r
\name{data.frame2nestedList}
\alias{data.frame2nestedList}
\alias{list2XML}
\alias{addAttributes}
\alias{getVarSetAttr}
\alias{appendToListKeepAttr}
\alias{extraxtAndSetDotDotAttributes}
\alias{setAttributes}
\title{Write biotic and acoustic XML files from data frames given a (preferably hierarcical) xsd (xml schema file).}
\usage{
data.frame2nestedList(
  x,
  pre = NULL,
  levelnames = NULL,
  rename = NULL,
  na.rm = FALSE,
  lastlevel = NULL
)

list2XML(node, sublist)

addAttributes(child, l)

getVarSetAttr(
  x,
  Attr,
  Var,
  pre = NULL,
  onlyfirst = TRUE,
  rename = NULL,
  na.rm = FALSE
)

appendToListKeepAttr(x, add)

extraxtAndSetDotDotAttributes(x)

setAttributes(x, att)
}
\arguments{
\item{x}{The data frame to write to an XML file, validated against the xsd. The data frame has one column per combination of variable and attribute, where the attributes are coded into the column names in the following manner: variableName..attributeName.attributeValue. If there are variables with identical names at different levels in the XMl hierarchy, the level (i.e., the name of the parent node) can be given in the column name by separation of a dot: variableName.level.}

\item{pre}{A nested list to add the nested list to (used in the recursion of \code{data.frame2nestedList}).}

\item{levelnames}{A vector of names of the levels of the resulting nested list.}

\item{rename}{A list of presentName = newName elements, which rename the columns in the data frame to be converted to a nested list.}

\item{na.rm}{Logical: If TRUE, remove missing values in the XML file (otherwise save these values as NA in the file).}

\item{lastlevel}{The level last processed in the recursion of \code{data.frame2nestedList}.}

\item{node}{The node to which to append the output from \code{list2XML} to.}

\item{sublist}{The list to append to the \code{node}.}

\item{child, l}{The XMl child to add attributes to, and the list from which the attributes are retrieved (used in \code{addAttributes}).}

\item{Attr, Var}{Inices of the columns which are variables and attributes at the current level.}

\item{onlyfirst}{Not sure why this is included...}

\item{add}{The list to add to the existinf list.}

\item{att}{A named list of attributes.}
}
\description{
\code{data.frame2xml} Converts a data frame to an XML object which can be written to file using XML::saveXML. \cr \cr
\code{data.frame2nestedList} Converts a data frame to a neste list (used by \code{data.frame2xml}). \cr \cr
\code{list2XML} Converts a list to an XML object. \cr \cr
\code{addAttributes} Function to add attributes. Used in \code{list2XML}. \cr \cr
\code{getVarSetAttr} Utility function for extracting a variable and setting attributes from a data frame. Used in \code{data.frame2nestedList}. \cr \cr
\code{appendToListKeepAttr} Function for appending a list 'add' to another list 'x', while keepeing the list names. Used in \code{data.frame2nestedList}. \cr \cr
\code{extraxtAndSetDotDotAttributes} Function for adding attributes stored in the names, which are ..AttrName.AttrValue. Used in \code{getVarSetAttr}. \cr \cr
\code{setAttributes} Function to set attributes 'att' to object 'x'. \cr \cr
}
\keyword{internal}
