% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_base.r
\name{data.frame2ParTable}
\alias{data.frame2ParTable}
\title{Get the Rstox version and the version of the Java library used by Rstox, on which StoX is built.}
\usage{
data.frame2ParTable(x)
}
\arguments{
\item{x}{A data.frame with the parameters as columns, such as data.frame(SpecCat=c("Torsk", "Sild", ""), Alpha=runif(3), Beta=runif(3), LMin=runif(3), LMax=runif(3)).}
}
\description{
Get the Rstox version and the version of the Java library used by Rstox, on which StoX is built.
}
\keyword{internal}
