% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RECA_plots.R
\name{diagnosticsCoverageRECA}
\alias{diagnosticsCoverageRECA}
\title{Paneled plots of cell coverage}
\usage{
diagnosticsCoverageRECA(stoxexport)
}
\description{
Paneled plots of cell coverage
}
\keyword{internal}
