% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_eca.r
\name{diagnosticsRECA}
\alias{diagnosticsRECA}
\title{Diagonostics RECA}
\usage{
diagnosticsRECA(projectName, verbose = T, format = "png", ...)
}
\arguments{
\item{projectName}{name of stox project}

\item{verbose}{logical, if TRUE info is written to stderr()}

\item{format}{function defining filtetype for plots, supports grDevices::pdf, grDevices::png, grDevices::jpeg, grDevices::tiff, grDevices::bmp}

\item{...}{parameters passed on to plot function and format}
}
\value{
list, with at least one named element 'filename', a vector of file-paths to generated plots.
}
\description{
Generate plots for diagnosis of RECA model configuration.
}
\details{
Plots are made conditional on problems. E.g. Fixed effects plot is not made, if all combinations of fixed effects were sampled.
}
