% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_reports.r
\name{distributeAbundance}
\alias{distributeAbundance}
\title{Impute unknown individual biological parameters from known values}
\usage{
distributeAbundance(i = NULL, abnd, seedV = NULL)
}
\arguments{
\item{i}{The index in the list of bootstrap iterations.}

\item{abnd}{Abundance matrix with individual data}

\item{seedV}{The seed vector for the random number generator, where element 'i' is picked out (this may seem strange, but is a consequence of the parallelability of the function, where 'i' is the primary parameter).}
}
\value{
Abundance matrix with imputed biological information
}
\description{
This function fills in holes in individual fish samples (also called imputation).
In cases where individuals are not aged, missing biological variables (e.g "weight","age","sex", and "specialstage") are sampled from 
fish in the same length group at the lowest imputation level possible.
imputeLevel = 0: no imputation, biological information exists
imputeLevel = 1: imputation at station level; biological information is selected at random from fish within station
imputeLevel = 2: imputation at strata level; no information available at station level, random selection within stratum
imputeLevel = 3: imputation at survey level; no information available at lower levels, random selection within suvey
imputeLevel = 99: no imputation, no biological information exists for this length group
}
\keyword{internal}
