% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RECA_plots.R
\name{formatPlot}
\alias{formatPlot}
\title{Saves plot to stox project location (getProjectPaths(projectname)$RReportDir) to be localised by getPlots in stox.}
\usage{
formatPlot(projectname, plotname, draw, format = "png", verbose = F, ...)
}
\arguments{
\item{projectname}{name of stox project}

\item{plotname}{filename for plot without format suffix}

\item{draw}{function for drawing the plot. Takes no arguments}

\item{format}{function defining plotting device and file suffix, supports grDevices::pdf, grDevices::png, grDevices::jpeg, grDevices::tiff, grDevices::bmp}

\item{verbose}{logical, if TRUE info is written to stderr()}

\item{...}{parameters to be passed on to format}
}
\description{
Saves plot to stox project location (getProjectPaths(projectname)$RReportDir) to be localised by getPlots in stox.
}
