% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RECA_plots.R
\name{get_eq_tail_ci}
\alias{get_eq_tail_ci}
\title{Constructs equal tailed credibility intervals. The probability mass above and below the interval is approximatelty the same.}
\usage{
get_eq_tail_ci(age, values, alpha)
}
\arguments{
\item{age}{agegroups integer vector}

\item{values}{matrix where age indexes rows, and columns index points in the posterior distribution}

\item{alpha}{value for the credibility intervals}
}
\value{
list with means, upper_ci and lower_ci all numeric vectors corresponding to age
}
\description{
Constructs equal tailed credibility intervals. The probability mass above and below the interval is approximatelty the same.
}
\keyword{internal}
