% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RECA_plots.R
\name{get_fixed_effects_landings}
\alias{get_fixed_effects_landings}
\title{Get aggregated landings for fixed effects}
\usage{
get_fixed_effects_landings(stoxexport)
}
\description{
Get aggregated landings for fixed effects
}
\keyword{internal}
