% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RECA_plots.R
\name{get_g_s_a_frame}
\alias{get_g_s_a_frame}
\title{get matrix of sample and landings from the subset of biotic that contains aged individuals}
\usage{
get_g_s_a_frame(eca)
}
\description{
get matrix of sample and landings from the subset of biotic that contains aged individuals
}
\keyword{internal}
