% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RECA_plots.R
\name{get_gta_landings}
\alias{get_gta_landings}
\title{Extracts all combinations of gear, temporal and spatial from landings}
\usage{
get_gta_landings(stoxexport)
}
\description{
Extracts all combinations of gear, temporal and spatial from landings
}
\keyword{internal}
