% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_eca.r
\name{getAgeGroupParamaters}
\alias{getAgeGroupParamaters}
\title{calculates catch matrix for given variable and unit}
\usage{
getAgeGroupParamaters(pred, plusgr = NULL)
}
\value{
data.frame with columns: age, meanLength, meanWeight, sd.of.meanLength, sd.of.meanWeight
}
\description{
calculates catch matrix for given variable and unit
}
\keyword{internal}
