% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_java.r
\name{getAvailableFunctions}
\alias{getAvailableFunctions}
\title{Get available StoX functions}
\usage{
getAvailableFunctions(type = "baseline")
}
\description{
This funcion gets available StoX functions.
}
\keyword{internal}

