% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_eca.r
\name{getCatchMatrix}
\alias{getCatchMatrix}
\title{calculates catch matrix for given variable and unit}
\usage{
getCatchMatrix(pred, var = "Abundance", unit = "ones")
}
\value{
list with three members (means, cv and caa_scaled)
}
\description{
calculates catch matrix for given variable and unit
}
\keyword{internal}
