% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_eca.r
\name{getCovariateMatrix}
\alias{getCovariateMatrix}
\title{Function for converting the covariates present in biotic or landing to integer:}
\usage{
getCovariateMatrix(data, covariateNames, allLevels)
}
\description{
Function for converting the covariates present in biotic or landing to integer:
}
\keyword{internal}
