% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_eca.r
\name{getCovariateValue}
\alias{getCovariateValue}
\title{Function used for extracting the correct covariate value from the inidces used in the covariate matrix passed to ECA:}
\usage{
getCovariateValue(index, eca, cov = "season", type = "biotic")
}
\description{
Function used for extracting the correct covariate value from the inidces used in the covariate matrix passed to ECA:
}
\keyword{internal}
