% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_java.r
\name{getDataFrame}
\alias{getDataFrame}
\alias{getDataFrame1}
\alias{getDataFrameAtLevel}
\alias{getProcessDataTableAsDataFrame}
\title{(Internal) Convert a storage object into a dataframe representation}
\usage{
getDataFrame(baseline, processName = NULL, functionName = NULL,
  level = NULL, drop = TRUE)

getDataFrame1(baseline, processName = NULL, functionName = NULL,
  level = NULL, drop = TRUE)

getDataFrameAtLevel(level, storage, data)

getProcessDataTableAsDataFrame(baseline, tableName)
}
\arguments{
\item{baseline}{A StoX baseline object}

\item{processName}{The name of the process to extract data by, such as "ReadBioticXML"}

\item{functionName}{The name of the function to extract data by, such as "ReadBioticXML"}

\item{level}{The level of tables for some datatypes like FishStation, which contains different levels, i.e., 1 for fishstation, 2 for sample and 3 for individuals. \code{getDataFrame1} selects the first level, and \code{getDataFrame} must be used to select a different level}

\item{drop}{Logical: if TRUE (defalut) drop the list if only one data frame is requested.}

\item{storage}{The StoX storage object}

\item{data}{The StoX data}

\item{tableName}{Supported processdata tables: TRAWLASSIGNMENT, DISTANCEASSIGNMENT, PSUASSIGNMENT, DISTANCEPSU, PSUSTRATUM, STRATUMPOLYGON}
}
\value{
A dataframe
}
\description{
\code{getDataFrame} converts a StoX storage object into a dataframe, and is used by the more user friendly \code{getDataFrame1} to convert a baseline object to dataframe. \cr \cr
\code{getProcessDataTableAsDataFrame} gets a joined table with meanNASC, psu, stratum, and area. Reads transect data, strata and area information from baseline Java object and merges them into one data frame.
}

