% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_eca.r
\name{getDataMatrixANDCovariateMatrix}
\alias{getDataMatrixANDCovariateMatrix}
\title{Function for extracting the DataMatrix for the given variable ("age" or "weight", length is requested in both):}
\usage{
getDataMatrixANDCovariateMatrix(eca, vars = c("age", "yearday"),
  ecaParameters)
}
\description{
Function for extracting the DataMatrix for the given variable ("age" or "weight", length is requested in both):
}
\keyword{internal}
