% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_eca.r
\name{getGlobalParameters}
\alias{getGlobalParameters}
\title{Function for extracting the GlobalParameters object}
\usage{
getGlobalParameters(biotic, resultdir, maxlength, minage, maxage,
  delta.age)
}
\description{
Function for extracting the GlobalParameters object
}
\keyword{internal}
