% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_eca.r
\name{getHardCoded}
\alias{getHardCoded}
\title{Function used for combining hard coded parameter values and user defeined parameter values}
\usage{
getHardCoded(info)
}
\description{
Function used for combining hard coded parameter values and user defeined parameter values
}
\keyword{internal}
