% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_eca.r
\name{getInfo}
\alias{getInfo}
\title{Function for extracting the CARNeighbours and info:}
\usage{
getInfo(eca, CovariateMatrix, modelSpecification = NULL)
}
\description{
Function for extracting the CARNeighbours and info:
}
\keyword{internal}
