% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_eca.r
\name{getLandings}
\alias{getLandings}
\title{Function for extracting the Landings object
Compiles landings objects aggreageted to the covariates in the models, and to the temporal resolution provided}
\usage{
getLandings(landing, AgeLength, WeightLength, landingresolution)
}
\description{
Function for extracting the Landings object
Compiles landings objects aggreageted to the covariates in the models, and to the temporal resolution provided
}
\keyword{internal}
