% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_eca.r
\name{getLengthGivenAge_Biotic}
\alias{getLengthGivenAge_Biotic}
\title{Function for converting to the input format required by ECA (this is the main function):}
\usage{
getLengthGivenAge_Biotic(
  eca,
  hatchDaySlashMonth,
  minage,
  maxage,
  onlyagestations = F
)
}
\arguments{
\item{onlyagestations}{If true, only hauls with some aged individiuals are used}
}
\description{
Function for converting to the input format required by ECA (this is the main function):
}
\keyword{internal}
