% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_eca.r
\name{getMidSeason}
\alias{getMidSeason}
\title{Function for setting getting appropriate value for the midSeason column in Landings}
\usage{
getMidSeason(x, tz = "UTC", format = "\%d/\%m/\%Y")
}
\arguments{
\item{x}{string representing a yearless date range as dd/mm-dd/mm or as a single day dd/mm}
}
\value{
the day number in the year for the given day, or the mean day number in the year for the endpoints of the range
}
\description{
Function for setting getting appropriate value for the midSeason column in Landings
}
\keyword{internal}
