% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_acoustic.r
\name{getNASCDistr}
\alias{getNASCDistr}
\title{Estimate mean NASC distribution}
\usage{
getNASCDistr(baseline, psuNASC, NASCDistr = "observed")
}
\arguments{
\item{baseline}{StoX Java baseline object.}

\item{psuNASC}{MeanNASC table from baseline.}

\item{NASCDistr}{Assumed distribution of mean NASC values, "normal", "lognormal", "gamma", or "weibull" ("normal" is default).}
}
\value{
list with mean(NASC) and var(NASC) per strata, global (stratified) mean, SE and RSE (Relative standard error) of NASC
}
\description{
Calculates mean and variance of NASC, based on Jolly & Hampton 1990 http://www.nrcresearchpress.com/doi/pdf/10.1139/f90-147
}
\examples{
projectName <- "Test_Rstox"
psuNASC <- getPSUNASC(projectName)
stratumNASC <- getNASCDistr(projectName, psuNASC=psuNASC, NASCDistr="normal")

}
