% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_base.r
\name{getParlist}
\alias{getParlist}
\alias{applyParlist}
\title{Merges a list and the following list elements given as "...".}
\usage{
getParlist(parlist = list(), ..., na.rm = TRUE)

applyParlist(parlist, fun)
}
\arguments{
\item{parlist}{List of parameters values specified as processName = list(parameter = value)}

\item{...}{Same as parlist, but can be specified separately (not in a list but as separate inputs).}

\item{na.rm}{Logical: If TRUE (default), remove NAs in the parlist.}

\item{fun}{The name of a function to which the parameters in parlist should be fit.}
}
\value{
Merged list of parameters, duplicates not removed.
}
\description{
Merges a list and the following list elements given as "...".
}
\keyword{internal}
