% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_base.r
\name{getProjectPaths}
\alias{getProjectDataEnv}
\alias{getProjectPaths}
\title{(Internal) Functions to return various paths and file names.}
\usage{
getProjectPaths(projectName = NULL, dir = NULL)

getProjectDataEnv(projectName)
}
\arguments{
\item{projectName}{The name or full path of the project, a baseline object (as returned from getBaseline() or runBaseline()), og a project object (as returned from open).}

\item{dir}{The directory holding the project(s).}
}
\value{
Various names and directories
}
\description{
\code{getProjectPaths} returns ta list of projectName, projectRoot, projectPath, RDataDir, RReportDir and projectXML. \cr \cr
\code{getProjectDataEnv} gets the project environment. \cr \cr
}

