% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_acoustic.r
\name{getResampledNASCDistr}
\alias{getResampledNASCDistr}
\title{Resample NASC distribution}
\usage{
getResampledNASCDistr(baseline, psuNASC, stratumNASC, parameters)
}
\arguments{
\item{baseline}{StoX Java baseline object.}

\item{psuNASC}{MeanNASC table from baseline.}

\item{stratumNASC}{Strata NASC estimates from getNASCDistr(baseline).}

\item{parameters}{Parameters set by user in Stox;
parameters$nboot: Number of bootstrap replicates
parameters$seed: The seed for the random number generator (used for reproducibility)}
}
\value{
Matrix of resampled strata NASC means
}
\description{
Generates resampled NASC distribution.
}
\examples{
projectName <- "Test_Rstox"
psuNASC <- getPSUNASC(projectName)
stratumNASC <- getNASCDistr(projectName, psuNASC=psuNASC, NASCDistr="normal")
resampledNASC <- getResampledNASCDistr(projectName, psuNASC=psuNASC, stratumNASC=stratumNASC, 
    parameters=list(seed=1, nboot=5))

}
