% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_base.r
\name{getRstoxVersion}
\alias{getRstoxVersion}
\title{Get the Rstox version and the version of the Java library used by Rstox, on which StoX is built.}
\usage{
getRstoxVersion()
}
\description{
Get the Rstox version and the version of the Java library used by Rstox, on which StoX is built.
}
\keyword{internal}
