% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_eca.r
\name{getWeightGivenLength_Biotic}
\alias{getWeightGivenLength_Biotic}
\title{Function for converting to the input format required by ECA (this is the main function):}
\usage{
getWeightGivenLength_Biotic(eca, ecaParameters)
}
\description{
Function for converting to the input format required by ECA (this is the main function):
}
\keyword{internal}
