% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_base.r
\name{hasAcousticData}
\alias{hasAcousticData}
\title{Identify whether acoustic data are present in the project}
\usage{
hasAcousticData(projectName)
}
\arguments{
\item{projectName}{The name or full path of the project, a baseline object (as returned from getBaseline() or runBaseline()), og a project object (as returned from open).}
}
\value{
list with mean and variance by stratum
}
\description{
Checks whether the process MeanNASC returns any output.
}

