% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_eca.r
\name{hasAgeInSample}
\alias{hasAgeInSample}
\title{Identifies whichindividuals are from catchsamples with some age readings.}
\usage{
hasAgeInSample(biotic)
}
\value{
logical vector identifying which individuals belong to a catchsample where age was sampled
}
\description{
Identifies whichindividuals are from catchsamples with some age readings.
}
\keyword{internal}
