% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_reports.r
\name{imputeByAge}
\alias{imputeByAge}
\title{Impute	missing individual data by age}
\usage{
imputeByAge(projectName, seed = 1, cores = 1, saveInd = TRUE)
}
\arguments{
\item{projectName}{The name or full path of the project, a baseline object (as returned from getBaseline() or runBaseline()), og a project object (as returned from open).}

\item{seed}{The seed for the random number generator (used for reproducibility)}

\item{cores}{An integer giving the number of cores to run the bootstrapping over.}

\item{saveInd}{Logical: if TRUE save the imputing indices.}
}
\value{
Updated list with imputed bootstrap results
}
\description{
Impute missing data within the bootstrap data object
}
\examples{
# Create the test project:
createProject("Test_Rstox", files=system.file("extdata", "Test_Rstox", package="Rstox"), ow=TRUE)
projectName <- "Test_Rstox"
bootstrap <- runBootstrap(projectName, nboot=10, acousticMethod=PSU~Stratum, bioticMethod=EDSU~Stratum)
# imputeByAge() fills in empty cells:
system.time(bootstrap_Acoustic_imputed <- imputeByAge(projectName))

}

