% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_base.r
\name{isProjectZipURL}
\alias{isProjectZipURL}
\title{Are the specified URLs acutally URLs to zip files?}
\usage{
isProjectZipURL(URL)
}
\arguments{
\item{URL}{The URL(s) of the zipped project(s).}
}
\description{
Are the specified URLs acutally URLs to zip files?
}
\keyword{internal}
