% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_base.r
\name{isURL}
\alias{isURL}
\title{Are the specified URLs acutally URLs to zip files?}
\usage{
isURL(URL, zip = FALSE)
}
\arguments{
\item{URL}{The URL(s) of the zipped project(s).}
}
\description{
Are the specified URLs acutally URLs to zip files?
}
\keyword{internal}
