% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_java.r
\name{JavaString2vector}
\alias{JavaString2vector}
\title{Convert a Java string to R vector}
\usage{
JavaString2vector(x)
}
\arguments{
\item{x}{A Java string imported from the StoX Java library.}
}
\description{
When reading some data from the StoX Java memory using toString(), the resulting string is converted to a string vector by this function.
}
\keyword{internal}
