% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_java.r
\name{jInt}
\alias{jBoolean}
\alias{jDouble}
\alias{jInt}
\title{Wrap basic R objects into Java objects}
\usage{
jInt(i)

jDouble(i)

jBoolean(i)
}
\arguments{
\item{i}{A single integer, double, or Boolean}
}
\value{
A Java integer, double, or Boolean
}
\description{
Wraps a value \code{i} into a Java integer, double, og Boolean for use in Java functions accessable by the package Rstox.
}
\keyword{internal}

