% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_boostrap.r
\name{linkPSU2Stratum}
\alias{linkPSU2Stratum}
\alias{checkNumPSUsInStratum}
\title{Bootstrap utils.}
\usage{
linkPSU2Stratum(x, projectName, psustratum = NULL, StratumArea = NULL,
  list.out = TRUE, ignore.case = TRUE, fill0 = FALSE)

checkNumPSUsInStratum(projectName)
}
\arguments{
\item{x}{A data frame with with one row per PSU, to which Stratum information should be added.}

\item{projectName}{The name or full path of the project, a baseline object (as returned from \code{\link{getBaseline}} or \code{\link{runBaseline}}, og a project object (as returned from \code{\link{openProject}}).}

\item{psustratum}{A data frame linking PSUs to stratum IDs. Should have the columns 'Stratum' and 'PSU' or 'SampleUnit'. Read from the project if missing.}

\item{StratumArea}{A data frame providing the area of each stratum. Should have the columns 'Area' and 'PolygonKey' or 'Stratum'. Read from the project if missing.}

\item{list.out}{Logical: If TRUE, split the data frame into a list per species category SpecCat.}

\item{ignore.case}{Logical: If TRUE, ingore case when splitting by species category SpecCat.}
}
\value{
A list with names "Stratum" and "Total" containing dataframes for each species with the following columns: Stratum, Mean, SD (standard deviation), Variance, CV (coefficient of variation), and lower and upper 90% confidence bounds.
}
\description{
\code{linkPSU2Stratum} Adds stratum information to a data frame with one row per PSU. Stratum ID, Area, number of PSUs and positive PSUs are added, and rows are generated for strata with non-positive density. \cr \cr
\code{checkNumPSUsInStratum} Checks for numner of PSUs >= 2 in each stratum, which is a requirement for variance estimation, and should possibly be a requirement for the bootstrapping when one of the methods is PSU~Stratum. \cr \cr
}
\keyword{internal}
