% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RECA_data_checks.R
\name{makeDataReportReca}
\alias{makeDataReportReca}
\title{Data report for RECA}
\usage{
makeDataReportReca(
  biotic,
  stationissuesfile,
  catchissuefile,
  imputationfile,
  verbose = T,
  covariates = NULL
)
}
\arguments{
\item{biotic}{sample data exported from stox in \code{\link[Rstox]{prepareRECA}}}

\item{catchissuefile}{file for writing catch sample issues}

\item{imputationfile}{file for writing report on issues that may need data massaging (imputation or estimation)}

\item{verbose}{If true, summary and file locations will be printed to stdout}

\item{stationussuesfile}{file for writing station issues}
}
\description{
Generates reports on data issues that might need to be addressed before running RECA
}
\details{
Checks data exported from stox for missing mandatory information, and issues that will invalidate common covariate-configurations
   issues with stations are written as a tab delimited fil to stationissuesfile
   issues with catch samples are written as a tab delimited fil to catchissuefile
   issues that can be solved with imputation or parameter estimation tools provided by stox are written to imputationfile
}
\keyword{internal}
