% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RECA_data_checks.R
\name{makeSampleHomogeneityReportRECA}
\alias{makeSampleHomogeneityReportRECA}
\title{Sample Homogeneity}
\usage{
makeSampleHomogeneityReportRECA(
  projectName,
  processName = "FilterBiotic",
  issuefile = NULL
)
}
\arguments{
\item{projectName}{name of stox project}

\item{processName}{process whose output is to be checked}

\item{issuefile}{filename for writing issues, if NULL, write to stderr()}
}
\description{
Checks selected parameters for homogeneity assumed by Reca.
 This is implemented to compensate variables that are not available for checking in the data exported by BioticCovData and LandingCovData.
}
\details{
runs a StoX-project up to the process specified by 'processName', which is assumed to putput biotic data.
 Writes an issue list if:
 \itemize{
  \item{There are more than one species used for the variable 'catchcategory'}
  \item{There are more than one value (including NA) used for the variable 'lengthmeasurement'}
  \item{Catch is fractioned (several samples for species at station) and product types are missing or differ between catch and sample weight ('catchproducttype' and 'sampleproducttype').}
 }
}
