% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_framework.r
\name{merge2}
\alias{merge2}
\title{Merge two data tables with all=TRUE and the specified columns and keys.}
\usage{
merge2(x, y, var = c("distance", "weight", "lengthsampleweight",
  "length", "lengthresolution"), keys = c("cruise", "serialnumber",
  "samplenumber", "SpecCat"), keys.out = FALSE)
}
\arguments{
\item{x, y}{Two data tables to be merged.}

\item{var}{A character vector of names of the columns to keep while merging.}

\item{keys}{A character vector of names of the columns to merge by (see the \code{by} argument in \code{\link[data.table]{merge}}).}

\item{keys.out}{Logical: If TRUE return a list with the keys used and the merged data.}
}
\value{
A merged data table.
}
\description{
Merges two data tables (see \code{\link[data.table]{data.table}}) with all=TRUE, while keeping only columns of the data tables with names intersecting \code{var}, and using the intersect of \code{keys} and the names of the data tables as the 'by' argument. By data tables
}
