% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_base.r
\name{moveToTrash}
\alias{moveToTrash}
\title{Move existing files to the trash directory in the directory of the first element (all file should be in the same directory).}
\usage{
moveToTrash(x)
}
\arguments{
\item{x}{A vector of file paths.}
}
\description{
Move existing files to the trash directory in the directory of the first element (all file should be in the same directory).
}
\keyword{internal}
