% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_NMD.R
\name{NMDcomposeSTS}
\alias{NMDcomposeSTS}
\title{Composing a single year Survey Time Series from scratch}
\usage{
NMDcomposeSTS(STSname, year, APIversion = NULL, DATAversion = NULL,
  ver = getRstoxDef("ver"), API = "http://tomcat7.imr.no:8080/apis/nmdapi",
  msg = FALSE)
}
\arguments{
\item{STSname}{Name of the survey time series. Must be a character vector.}

\item{year}{Year of the survey time searies. Must be a character vector.}

\item{APIversion}{Placeholder, not used yet (see below TODO).}

\item{DATAversion}{Placeholder, not used yet (see below TODO).}
}
\value{
A character vector containing the newly created StoX project directory
}
\description{
\code{NMDcomposeSTS} returns the created StoX project path.
}
\examples{
\dontrun{
NMDcomposeSTS("Barents Sea Beaked redfish and Sebastes sp in Subareas
I and II bottom trawl index in winter", "2017")
}
}
