% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_NMD.R
\name{NMDdecode}
\alias{NMDdecode}
\title{Encodes and decodes NMD API strings.}
\usage{
NMDdecode(URL)
}
\arguments{
\item{URL}{An URL.}
}
\description{
\code{getNMDinfo} converts, prints and optionally returns NMD reference information given a search string to the reference information. Used in StoX.URL(). \cr 
\code{getNMDdata} downloads data from specific cruises, cruise series ot survey time series from NMD. \cr 
\code{downloadXML} downloads xml data from an API, parses the xml data, and converts to a list (the latter is time consuming).
}
\keyword{internal}

