% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RECA_plots.R
\name{plot_catch_at_age_ci}
\alias{plot_catch_at_age_ci}
\title{Plot equal tailed credible intervals for a catch by age prediction}
\usage{
plot_catch_at_age_ci(pred, var, unit, alpha = 0.1, xlab = "age",
  ylab = paste("posterior catch", unit), ...)
}
\arguments{
\item{pred}{RECA prediction object as returned by eca::eca.predict}

\item{var}{A key string indicating the variable to plot. ´Abundance´ and ´Weight´ is implemented.}

\item{unit}{A unit key string indicating the unit (see getPlottingUnit()$definitions$unlist.units for available key strings)}
}
\description{
Plot equal tailed credible intervals for a catch by age prediction
}
\keyword{internal}
