% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RECA_plots.R
\name{plot_catch_fractions}
\alias{plot_catch_fractions}
\title{Plots composition in catchsamples of parameters that determines what kind fraction of catches or landings are sampled}
\usage{
plot_catch_fractions(biotic, title = "sampling point",
  xlab = "# catch samples", allname = "Unsorted", landname = "Landed",
  discname = "Not landed", allcol = "#fee8c8", disccol = "#fdbb84",
  landcol = "#e34a33", unkowncol = "white", blankcode = default_blankcode,
  barplot = F)
}
\arguments{
\item{title}{title for plot}

\item{xlab}{label for x axis}

\item{allname}{name to use for samples where all catch was sampled}

\item{landname}{name to use for samples where only landed fraction was sampled}

\item{discname}{name to use for discared fraction of cathc when samples where taken after sorting}

\item{allcol}{color to use for samples where all catch was sampled}

\item{landcol}{color to use for samples where only landed fraction was sampled}

\item{blankcode}{code to use when fraction is not coded in data.}

\item{barplot}{if T barplot is plotted in stead of pie chart}

\item{disccolcolor}{to use for samples where only discared fraction was sampled}

\item{unkwoncol}{color to use when sampled fraction is not coded in data.}
}
\description{
Plots composition in catchsamples of parameters that determines what kind fraction of catches or landings are sampled
}
\keyword{internal}
