% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RECA_plots.R
\name{plot_cell_coverage}
\alias{plot_cell_coverage}
\title{Plot aggregated landings for different kind of sampling level in cells}
\usage{
plot_cell_coverage(eca, xlab = "sample clusteredness",
  ylab = "Fraction landed (weight-\%)",
  titletext = "Coverage w age\\ncells (gear/temp/spatial)",
  colgood = default_color_good, colok = default_color_ok,
  colbarely = default_color_barely, colbad = default_color_bad,
  colempty = default_color_empty, colwrong = default_color_wrong,
  gooddesc = "> 1 vessel", okdesc = "> 1 catch", barelydesc = "> 0 catch",
  baddesc = "0 samples")
}
\description{
Plot aggregated landings for different kind of sampling level in cells
}
\keyword{internal}
