% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RECA_plots.R
\name{plot_fixed_effect_coverage}
\alias{plot_fixed_effect_coverage}
\title{Plot table of coverage for fixed effects}
\usage{
plot_fixed_effect_coverage(stoxexport, indparameters = c("age"),
  titletext = "Samples for fixed effects", okcol = default_color_ok,
  wrongcol = default_color_wrong, undersampledcol = default_color_empty)
}
\arguments{
\item{stoxexport}{as returned from \code{\link{baseline2eca}}}

\item{indparameter}{the parameters for which data needs to be available}
}
\description{
Plot table of coverage for fixed effects
}
\keyword{internal}
