% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RECA_plots.R
\name{plot_length_at_age}
\alias{plot_length_at_age}
\title{Plots mean length at age and sample range}
\usage{
plot_length_at_age(biotic, pred, xlab = "age", alpha = 0.01,
  ylab = paste("length cm"), ...)
}
\arguments{
\item{biotic}{indiviuals as exported by stox}

\item{pred}{RECA prediction object as returned by eca::eca.predict}
}
\description{
Plots mean length at age and sample range
}
\keyword{internal}
