% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RECA_plots.R
\name{plot_mission_types}
\alias{plot_mission_types}
\title{Composition of mission types in the data}
\usage{
plot_mission_types(biotic, title = "mission types\\n# stations",
  blankcode = default_blankcode, pal = c("#e41a1c", "#377eb8", "#4daf4a",
  "#984ea3", "#ff7f00", "#ffff33", "#a65628", "#f781bf", "#999999"))
}
\description{
Composition of mission types in the data
}
\keyword{internal}
