% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RECA_plots.R
\name{plot_pred_box}
\alias{plot_pred_box}
\title{Plot catch by age prediction as boxplots}
\usage{
plot_pred_box(pred, var, unit, xlab = "age", ylab = paste("posterior catch",
  unit), ...)
}
\arguments{
\item{pred}{RECA prediction object as returned by eca::eca.predict}

\item{var}{A key string indicating the variable to plot. 'Abundance' and 'Weight' is implemented.}

\item{unit}{A unit key string indicating the unit (see getPlottingUnit()$definitions$unlist.units for available key strings)}
}
\description{
Plot catch by age prediction as boxplots
}
\keyword{internal}
