% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RECA_plots.R
\name{plot_sample_types}
\alias{plot_sample_types}
\title{Composition of catch sample types in data}
\usage{
plot_sample_types(biotic, title = "sample types",
  xlab = "# catch samples", blankcode = default_blankcode,
  cex.names = 0.8)
}
\arguments{
\item{title}{title for plot}

\item{xlab}{label for x axis}

\item{blankcode}{code for NA / not registered}

\item{cex.names}{expansion factor for bar labels}
}
\description{
Composition of catch sample types in data
}
\keyword{internal}
