% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RECA_plots.R
\name{plot_station_types}
\alias{plot_station_types}
\title{Composition of station types (Stasjons: stasjonstype) in data}
\usage{
plot_station_types(biotic, title = "station types",
  xlab = "# stations", blankcode = default_blankcode,
  cex.names = 0.8)
}
\arguments{
\item{title}{title for plot}

\item{xlab}{label for x axis}

\item{blankcode}{code for NA / not registered}

\item{cex.names}{expansion factor for bar labels}
}
\description{
Composition of station types (Stasjons: stasjonstype) in data
}
\keyword{internal}
