% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RECA_plots.R
\name{plot_weight_at_age}
\alias{plot_weight_at_age}
\title{Plots mean weight at age and sample range}
\usage{
plot_weight_at_age(biotic, pred, unit, alpha = 0.01, xlab = "age",
  ylab = paste("ind. weight", unit), ...)
}
\arguments{
\item{biotic}{indiviuals as exported by stox}

\item{pred}{RECA prediction object as returned by eca::eca.predict}
}
\description{
Plots mean weight at age and sample range
}
\keyword{internal}
