% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_reports.r
\name{plotAbundance}
\alias{plotAbundance}
\title{Plot abundance results to file (generic)}
\usage{
plotAbundance(projectName, var = "Abundance", unit = NULL,
  baseunit = NULL, grp1 = "age", grp2 = NULL, xlab = NULL,
  ylab = NULL, main = "", format = "png", maxcv = 1, ...)
}
\arguments{
\item{projectName}{The name or full path of the project, a baseline object (as returned from getBaseline() or runBaseline()), og a project object (as returned from open).}

\item{var}{A key string indicating the variable to plot (see getPlottingUnit()$defaults$Rstox_var for available values).}

\item{unit}{A unit key string indicating the unit, or alternatively a numeric value giving the scaling factor (run getPlottingUnit() to see available values).}

\item{baseunit}{The unit used in the data.}

\item{grp1}{Variable used to group results, e.g. "age", "LenGrp", "sex"}

\item{grp2}{An optional second grouping variable}

\item{xlab}{The label to user for the x axis, with default depending on data plotted,}

\item{ylab}{The label to user for the y axis, with default depending on data plotted.}

\item{main}{Main title for plot (text)}

\item{format}{The file format of the saved plot, given as a string naming the function to use for saving the plot (such as bmp, jpeg, png, tiff), with \code{filename} as its first argument. Arguments fo the functions are given as \code{...}. Dimensions are defaulted to width=800, height=600.}

\item{maxcv}{The maximum cv in the plot. Use Inf to indicate the maximum cv of the data.}

\item{...}{Parameters passed on from other functions. Includes \code{numberscale}, which is kept for compability with older versions. Please use 'unit' instead. (Scale results with e.g. 1000 or 1000000).}
}
\value{
Plot saved to file and abundance table printed
}
\description{
Plots boxplot of bootstrap results together with Coefficient of Variation (CV).
Prints summary table and plot is exported to a tif- or png-file.
}
\examples{
projectName <- "Test_Rstox"
plotAbundance(projectName, grp1="age")
plotAbundance(projectName, grp1="age", unit=1)

}
