% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_plots.R
\name{plotFishStation}
\alias{plotFishStation}
\alias{plotLonLat}
\alias{zoom_xlim_ylim}
\alias{zoom_lon_lat}
\alias{getFishStationOneSpecies}
\title{Plot fish stations by species}
\usage{
plotFishStation(projectName, species = "torsk", proc = "FilterBiotic",
  size = 1, col = 1, shape = 16, alpha = 1, jitter = FALSE,
  zoom = 1, offset = c(0.5, 0.5))

plotLonLat(x, lon = "lon", lat = "lat", species = "species",
  size = 1, zoom = 1, offset = c(0.5, 0.5), col = 1, shape = 16,
  alpha = 1, jitter = FALSE)

zoom_xlim_ylim(xlim, ylim, zoom = 1, offset = c(0.5, 0.5))

zoom_lon_lat(x, lon = "lon", lat = "lat", zoom = 1, offset = c(0.5,
  0.5))

getFishStationOneSpecies(species = "torsk", projectName,
  proc = "FilterBiotic", msg = FALSE)
}
\arguments{
\item{projectName}{The name or full path of the project, a baseline object (as returned from \code{getBaseline} or \code{runBaseline}), og a project object (as returned from \code{openProject}). For \code{createProject}, \code{projectName}=NULL (the default) returns available templates, and for \code{openProject}, zeros length \code{projectName} returns all StoX projects in the default workspace either given as a vector of full paths, or, in the case projectName is an empty list, a list of names of StoX projects located in the default workspace and sub directories. Projects locataed in sub directories of the default workspace can be given by the relative path, or are searched for by name.}

\item{species}{A vector of the species for which stations will be plotted, given in the current version as the norwegian name ('noname'). there can be several species for each station, which will lead to several points in the same location (see \code{jitter}).}

\item{proc}{The process from which to get the fish stations (an object including the sting "FishStation").}

\item{size}{The size of the points, equal for all points in the current version.}

\item{col, shape}{The color/shape of the fish stations. If TRUE, use ggplot colors/shapes for each species.}

\item{alpha}{Transparency in the range [0,1].}

\item{jitter}{A vector of 2 or 3 elements, where the first two are the jitter scales in the x (longitude) and y (latitude) direction, respectively (), and the third element is the seed for the jittering. Use jittering with caution, since it shifts the position of the fish stations!}

\item{zoom, offset}{Used for zooming the plot. A value of zoom = 1 indicates no zoom, values samller than 1 zooms in and values larger than one zooms out. The \code{offest} sets the center of the zoom relative to the limits of the original plot, where 0.5 (default) indicates the center of the plot.}
}
\value{
A ggplot
}
\description{
This function plots the fish stations of the biotic data of a StoX project. The stations can be plotted with colors indicating the species specified by the parameter \code{species}. For stations with more than one of these species present, the last is plotted last, and this will overlay the other species in the station. To avoid this, there is an option \code{jitter}.
}
\details{
These are the details
\deqn{\sum_{i} \hat{f} (x_i) \frac{b_{i+1}} {b_i}  = 1}
End of the details
}
\examples{
projectName <- "Test_Rstox"
# Show the species sampled during the cruise:
g <- getBaseline(projectName, input=FALSE, proc="ReadBioticXML")
speciesFrequency <- sort(table(g$ReadBioticXML_BioticData_CatchSample.txt$noname), decreasing=TRUE)
species <- names(speciesFrequency)

# Plot the 6 most frequent species:
plotFishStation(projectName, species=species[1:6], col=TRUE, size=2)
# Add jitter to the plot, resolving stations with multiple species. 
# CAUTION: This adds randmoness to the fish station positions!:
plotFishStation(projectName, species=species[1:6], col=TRUE, size=2, jitter=c(0.2, 0.2, 1))

}
