% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_eca.r
\name{plotRECA}
\alias{plotRECA}
\title{Produces plots for R report for RECA. Fails silently on errors.}
\usage{
plotRECA(projectName, ...)
}
\arguments{
\item{projectName}{name of stox project}
}
\value{
list, with at least one named element 'filename', a vector of file-paths to generated plots.
}
\description{
Produces plots for R report for RECA. Fails silently on errors.
}
