% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_eca.r
\name{plotRECAresults}
\alias{plotRECAresults}
\title{Generates plots and reports from RECA prediction}
\usage{
plotRECAresults(projectName, verbose = F, format = "png", ...)
}
\arguments{
\item{projectName}{name of stox project}

\item{verbose}{logical, if TRUE info is written to stderr()}

\item{format}{function defining filtetype for plots, supports grDevices::pdf, grDevices::png, grDevices::jpeg, grDevices::tiff, grDevices::bmp}

\item{...}{parameters passed on plot function and format}
}
\description{
Generates plots and reports from RECA prediction
}
