% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RECA_plots.R
\name{plotSampleCompositionRECA}
\alias{plotSampleCompositionRECA}
\title{Panelled plot of sample composition wrp potentially problematic heterogenety}
\usage{
plotSampleCompositionRECA(biotic, ...)
}
\arguments{
\item{projectName}{name of stox project}
}
\description{
Panelled plot of sample composition wrp potentially problematic heterogenety
}
\keyword{internal}
