% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_eca.r
\name{plotSamplingOverview}
\alias{plotSamplingOverview}
\title{plotSamplingOverview}
\usage{
plotSamplingOverview(projectName, verbose = T, format = "png", ...)
}
\arguments{
\item{projectName}{name of stox project}

\item{verbose}{logical, if TRUE info is written to stderr()}

\item{format}{function defining filtetype for plots, supports grDevices::pdf, grDevices::png, grDevices::jpeg, grDevices::tiff, grDevices::bmp}

\item{...}{parameters passed on to plot function and format}
}
\value{
list, with at least one named element 'filename', a vector of file-paths to generated plots.
}
\description{
Generate plots to show composition of samples wrp activity in fisheries
}
\details{
Compares sampling effort to fisheries along covariates selected in the model, and along some standard covariate choices if available (gear, temporal and spatial). Plots compositions of samples with respect to some important variables informative of sampling heterogenety
}
