% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_base.r
\name{pointToStoXFiles}
\alias{pointToStoXFiles}
\title{Link to the files in a stox project}
\usage{
pointToStoXFiles(projectName, files = NULL)
}
\arguments{
\item{projectName}{The name or full path of the project, a baseline object (as returned from \code{\link{getBaseline}} or \code{\link{runBaseline}}, og a project object (as returned from \code{\link{openProject}}).}

\item{files}{A list with elements named "acoustic", "biotic", "landing", "process" (holding the project.xml file) or other implemented types of data to be copied to the project (available data types are stored in StoX_data_types in the environment "RstoxEnv". Get these by get("StoX_data_types", envir=get("RstoxEnv"))). These could be given as directories, in which case all files in those directories are copied, or as URLs. If given as a single path to a directory holding sub-directories with names "acoustic", "biotic", "landing", "process" or other implemented types of data, the files are copied from these directories. If files has length 0 (default), the files present in the project directory are used, if already existing (requires to answer "y" when asked to overwrite the project if ow=NULL, or alternatively to set ow=TRUE).}
}
\value{
A project object
}
\description{
Updates a project with the files located in the "input" directory. Used in updateProject().
}
\keyword{internal}
