% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_spatial.r
\name{polyArea}
\alias{polyArea}
\title{Polygon area}
\usage{
polyArea(wkt)
}
\arguments{
\item{wkt}{String like "MULTIPOLYGON(((4.9 57.1, 5.00 57.23, 4.0 58.9)))}
}
\value{
Area in nmi squared
}
\description{
Calculate area of wkt polygon with az.eq.are projection / adapted origo.
}
\examples{
poly <- cbind(c(4.9, 5.00, 4.0, 4.9), c(57.1, 57.23, 58.9, 57.1))
plot(NULL, xlim=range(poly[,1]), ylim=range(poly[,2]), xlab="Longitude", ylab="Latitude")
polygon(poly, col="green")
polyArea(poly)

}

