% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_eca.r
\name{prepareRECA}
\alias{prepareRECA}
\title{Prepare data for RECA}
\usage{
prepareRECA(
  projectName,
  resultdir = NULL,
  minage = 1,
  maxage = 20,
  delta.age = 0.001,
  maxlength = NULL,
  hatchDaySlashMonth = "01/01",
  temporalresolution = 92,
  overwrite = T,
  agedstationsonly = F,
  landingAdjuster = NULL
)
}
\arguments{
\item{projectName}{name of stox project}

\item{resultdir}{location where R-ECA will store temporal files. Defaults (if null) to a subdirectory of getProjectPaths(projectName)$RDataDir called `reca` whcih will be created if it does not already exist}

\item{minage}{see specification for GlobalParameters in \code{\link[Reca]{eca.estimate}}.}

\item{maxage}{see specification for GlobalParameters in \code{\link[Reca]{eca.estimate}}}

\item{delta.age}{see specification for GlobalParameters in \code{\link[Reca]{eca.estimate}}}

\item{maxlength}{maximum length of fish in the data set in cm. If null the value will be extracted from the data.}

\item{hatchDaySlashMonth}{reference day for assumed spawning time of fish, formatted as day / month. Used to estimate fractional age of fish.}

\item{temporalresolution}{temporal resolution for the aggregated landings in days (used to set midSeason the Landings object for \code{\link[Reca]{eca.predict}})}

\item{overwrite}{logical, if true, projectData for prepareRECA and runRECA will be nulled before running, and resultdir will be cleaned of any existing output files located in subdirectories cfiles and resfiles.}

\item{agedstationsonly}{logical, if true, only hauls with some aged individuals will be used for the age model. This does not affect the weight-length model}

\item{landingAdjuster}{optional function to manipulate landings before conversion. Used when running from R, for setting covariates not supported by the StoX pipeline, adjusting with logbook records, etc. input and output format for this funcition can be inspected from the output of \code{\link[Restox]{prepareRECA}} (loadProjectData(projectName)$prepareRECA$StoxExport$landing)}
}
\description{
Convert data to exported from stox to eca format. Save results to project data 'prepareRECA'
}
\details{
Most parameters to this funciton are set as named members of a list which is passed as argument GlobalParameters to \code{\link[Reca]{eca.estimate}}
   The parameters minage and maxage define the range of ages that are considered possible in the model. Because R-ECA integrates weight and length measurements, and allows for modelling errors in age determination, predicted ages might fall outside the age range in samples. minage and maxage should be set with this in mind.
   
   In order to provide ad-hoc support for manipulations of landings data not provided by StoX
}
