% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_datras.R
\name{prepDatras}
\alias{prepDatras}
\title{Generate DATRAS output for a specific StoX project  
or user specified cruise number and ship name. 
NOTE: There are some differences between the output of this function 
and the previous submissions in the HH file due to lack of trawl sensor and CTD data:
(HydroStNo, Netopening, DoorSpread, WingSpread) (SurTemp BotTemp SurSal BotSal)}
\usage{
prepDatras(datrasProject)
}
\arguments{
\item{datrasProject}{The name or full path of the project, a baseline object (as returned from \code{getBaseline} or \code{runBaseline}), or a project object (as returned from \code{openProject}). Projects located in sub directories of the default workspace can be given by the relative path, or are searched for by name.}
}
\description{
Generate DATRAS output for a specific StoX project  
or user specified cruise number and ship name. 
NOTE: There are some differences between the output of this function 
and the previous submissions in the HH file due to lack of trawl sensor and CTD data:
(HydroStNo, Netopening, DoorSpread, WingSpread) (SurTemp BotTemp SurSal BotSal)
}
\examples{
# Process existing project
datras <- prepDatras("NMD_CruiseNumber_2018102_ShipName_G.O.Sars")
head(datras$outputData$DATRASConvert$DATRASConvert_BioticData_HH.txt)
head(datras$outputData$DATRASConvert$DATRASConvert_BioticData_HL.txt)

}
