% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_datras.R
\name{processDatras}
\alias{processDatras}
\title{Generate DATRAS output for a specific StoX project  
or user specified cruise number and ship name. 
NOTE: There are some differences between the output of this function 
and the previous submissions in the HH file due to lack of trawl sensor and CTD data:
(HydroStNo, Netopening, DoorSpread, WingSpread) (SurTemp BotTemp SurSal BotSal)}
\usage{
processDatras(datrasProject, cruiseNumber = NULL, cruiseShip = NULL,
  ver = list(biotic = 3, API = list(biotic = 3)),
  server = "http://tomcat7.imr.no:8080/apis/nmdapi", snapshot = TRUE,
  overwrite = FALSE)
}
\arguments{
\item{datrasProject}{The name or full path of the project, a baseline object (as returned from \code{getBaseline} or \code{runBaseline}), og a project object (as returned from \code{openProject}). For \code{createProject}, \code{projectName}=NULL (the default) returns available templates, and for \code{openProject}, zeros length \code{projectName} returns all StoX projects in the default workspace either given as a vector of full paths, or, in the case projectName is an empty list, a list of names of StoX projects located in the default workspace and sub directories. Projects locataed in sub directories of the default workspace can be given by the relative path, or are searched for by name.}

\item{cruiseNumber}{If not NULL, create a new StoX project for DATRAS conversion with a specific cruise number, must be use together with cruiseShip.}

\item{cruiseShip}{If not NULL, create a new StoX project for DATRAS conversion with a specific cruise name, must be use together with cruiseNumber.}

\item{ver}{The version of the APIs and data, given as a list such as that returned by getRstoxDef("ver"). To use version 1 of the API (the only one available prior to Rstox 1.10) use getRstoxDef("ver", API=list(biotic=1, reference=1)), or simply ver = 1. Giving \code{ver} as a single integer imposes this value on all API versions. To set API versions to 1 and at the same time set the version of the data, use e.g. ver = list(API=1, biotic=1.4). Only valid when creating a new StoX project for DATRAS conversion.}

\item{server}{The path of the server. Only valid when creating a new StoX project for DATRAS conversion.}

\item{snapshot}{Specifies whether the biotic data is downloaded from snapshot. Only valid when creating a new StoX project for DATRAS conversion.}

\item{overwrite}{Specifies whether to ovewrite existing project (if any): If TRUE, overwrite; if FALSE, do not overwrite. Only valid when creating a new StoX project for DATRAS conversion.}
}
\description{
Generate DATRAS output for a specific StoX project  
or user specified cruise number and ship name. 
NOTE: There are some differences between the output of this function 
and the previous submissions in the HH file due to lack of trawl sensor and CTD data:
(HydroStNo, Netopening, DoorSpread, WingSpread) (SurTemp BotTemp SurSal BotSal)
}
\examples{
# Process existing project
datras <- processDatras("IBTS-Q1_2016")
head(datras$HH)
head(datras$HH)
# Create new project

}
