% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_base.r
\name{readBaselineFiles}
\alias{readBaselineFiles}
\title{Function for reading output files from StoX (Baseline or Baseline report)}
\usage{
readBaselineFiles(x)
}
\arguments{
\item{x}{The files to read.}
}
\description{
Function for reading output files from StoX (Baseline or Baseline report)
}
\keyword{internal}
