% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logbooks.R
\name{readLstFile}
\alias{readLstFile}
\title{Parses logbooks (ERS)}
\usage{
readLstFile(file, encoding = "latin1")
}
\arguments{
\item{file}{path to file}

\item{encoding}{encoding for 'file'}
}
\value{
data.table() with logbooks, columns described in details.
}
\description{
Parses electronic logbooks (ERS) from tabular format .lst
}
\details{
The format is a ;-separated format encoding aggregated ERS records (logbooks).
 It is based on files provided to IMR on a regular basis from FDIR.
 Column headers are in Norwegian.
 
 Columns:
 \describe{
   \item{FAAR}{Year of catch}
   \item{REGM}{License number of vessel}
   \item{RKAL}{Radio call signal}
   \item{FM}{Month of catch}
   \item{FD}{Day of catch}
   \item{DBNR}{Logbook number}
   \item{TUR}{Trip id}
   \item{AM}{Month of depature for trip}
   \item{AD}{First day of departure for trip}
   \item{AH}{Port of departure}
   \item{LM}{Month of landing of catch}
   \item{LD}{Day of landing of catch}
   \item{LH}{Port of landing of catch}
   \item{RE}{Gear (main gear for the day of catch)}
   \item{MA}{Mesh-size of gear}
   \item{HA}{Number of hauls / sets for the day of catch}
   \item{VAR}{Total duration of hauls for the day of catch}
   \item{OMRA}{Area code (ICES, NAFO etc)}
   \item{OKSO}{Economic zone}
   \item{HO}{Main area (defined by Norwegian directorate of Fisheries)}
   \item{LO}{Location (within main area)}
   \item{LENG}{Vessel length in meters}
   \item{BTON}{Gross tonnage of vessel}
   \item{TENH}{Tonnage of vessel}
   \item{HEST}{Enigine horsepower for vessel}
   \item{FISK}{Code for type of fish (species). Code list provided by Norwegian directorate of Fisheries}
   \item{VEKT}{Love weight in kg}
 }
}
