% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_reports.r
\name{reportAbundance}
\alias{reportAbundance}
\alias{reportAbundanceAtLevel}
\title{Calculate a summary of the bootstrap iterations (possibly after imputing unknown ages).}
\usage{
reportAbundance(projectName, var = "Abundance", unit = NULL,
  baseunit = NULL, grp1 = "age", grp2 = NULL, numberscale = 1e+06,
  plotOutput = FALSE, write = FALSE)

reportAbundanceAtLevel(projectName, var = "Abundance", unit = NULL,
  baseunit = NULL, level = "bootstrapImpute", grp1 = "age", grp2 = NULL,
  numberscale = 1e+06, plotOutput = FALSE, write = FALSE)
}
\arguments{
\item{projectName}{The name or full path of the project, a baseline object (as returned from getBaseline() or runBaseline()), og a project object (as returned from open).}

\item{var}{A key string indicating the variable to plot (see getPlottingUnit()$defaults$Rstox_var for available values).}

\item{unit}{A unit key string indicating the unit, or alternatively a numeric value giving the scaling factor (run getPlottingUnit() to see available values).}

\item{baseunit}{The unit used in the data.}

\item{grp1}{Variable used to group results, e.g. "age", "LenGrp", "sex"}

\item{grp2}{An optional second grouping variable}

\item{numberscale}{Kept for compability with older versions. Use 'unit' instead. (Scale results with e.g. 1000 or 1000000).}

\item{plotOutput}{Logical: if TRUE return a list of the recuired data in the function plotAbundance(). Otherwise, only the abundance data frame is returned.}

\item{write}{Logical: if TRUE write the data to a tab-separated file.}

\item{level}{A string naming the process level (see getRstoxEnv()$processLevels for available levels).}
}
\value{
A data frame of the abundance in sumary per grp2 and grp1 if plotOutput=FALSE, and a list holding this object (keeping "-" for missing values and not ordering) and other objects needed by plotAbundance().
}
\description{
This function is used in the plotting function plotAbundance().
}
\examples{
projectName <- "Test_Rstox"
reportAbundance(projectName, grp1=NULL, grp2=NULL)
reportAbundance(projectName, grp1="age", grp2=NULL)
reportAbundance(projectName, var="weight", grp1="age", grp2=NULL)
reportAbundance(projectName, grp1="age", grp2="Stratum")
reportAbundance(projectName, grp1="age", grp2="Stratum")

}
