% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_eca.r
\name{reportRECA}
\alias{reportRECA}
\title{Report RECA.}
\usage{
reportRECA(projectName, var = "Abundance", unit = "ones", write = T,
  ...)
}
\arguments{
\item{projectName}{name of stox project}

\item{var}{Variable to extract. Allows for Abundance, Count or Weight}

\item{unit}{Unit for extracted variable. See \code{\link{getPlottingUnit}}}

\item{write}{logical determining if report is written to files. If false the function return immidiatly with NULL.}

\item{...}{arguments passed to \code{\link{saveCatchMatrix}} and \code{\link{writeRecaConfiguration}}}
}
\value{
list, with at least one named element 'filename', a vector of file-paths to generated plots.
}
\description{
Produces reports for for RECA. Fails silently on errors.
}
\details{
Exports a tab separated file with means of catch at age (produced by \code{\link{saveCatchMatrix}}), one for the posterior distribution of catch at age (produced by \code{\link{saveCatchMatrix}}), and a file summarizing the model configuration (produced by \code{\link{writeRecaConfiguration}})
   This function is intended for use with \code{\link{getReports}}, and contain parameters like write in order to adhere to that contract.
}
