% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_base.r
\name{rm.na}
\alias{rm.na}
\title{Small function for removing missing values if requested}
\usage{
rm.na(x, na.rm = TRUE)
}
\arguments{
\item{x}{The files to read.}

\item{na.rm}{Logical: If FALSE do not remove the NAs.}
}
\description{
Small function for removing missing values if requested
}
\keyword{internal}
