% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{Rstox-package}
\alias{Rstox}
\alias{Rstox-package}
\title{Running StoX functionality independently in R
 The package Rstox contains most of the functionality of the stock assesment utility StoX, which is an open source application fo acoustic-trawl and swept-area survey estimation. #'}
\description{
The package Rstox contains most of the functionality of the stock assesment utility StoX, which is an open source application fo acoustic-trawl and swept-area survey estimation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Sea2Data/Rstox}
  \item Report bugs at \url{https://github.com/Sea2Data/Rstox/issues}
}

}
\author{
\strong{Maintainer}: Arne Johannes Holmin \email{arnejh@hi.no}

Authors:
\itemize{
  \item Ibrahim Umar
  \item Edvin Fuglebakk
  \item Aasmund Skaalevik
  \item Espen Johnsen
  \item Atle Totland
  \item Gjert Endre Dingsoer
}

}
