% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_base.r
\name{sampleSorted}
\alias{sampleSorted}
\alias{getSeeds}
\title{Sample a vector after sorting the vector and applying the seed}
\usage{
sampleSorted(x, size, lx = NULL, seed = 0, replace = TRUE,
  sorted = TRUE)

getSeeds()
}
\arguments{
\item{x}{A vector or a single integer.}

\item{size}{A non-negative integer giving the number of items to choose.}

\item{lx}{The (optional) length of\code{x}.}

\item{seed}{The seed to apply before sampling.}

\item{replace}{Should sampling be with replacement?}

\item{sorted}{Should \code{x} be sorted prior to sampling?}
}
\description{
Sample a vector after sorting the vector and applying the seed
}
\keyword{internal}
