% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_eca.r
\name{saveAgeGroupParameters}
\alias{saveAgeGroupParameters}
\title{Save age group parameters (mean length and weight)}
\usage{
saveAgeGroupParameters(pred, filename, main = "", plusgr = NULL)
}
\arguments{
\item{pred}{as returned by \code{\link[Reca]{eca.predict}}}

\item{filename}{name of file to save to.}

\item{main}{Title for the analysis, to be included as comment in saved file (e.g. species and year)}

\item{plusgr}{Lower age in plusgr for tabulation. If NULL plusgr is not used.}
}
\description{
Write age group parameters predicted by \code{\link[Reca]{eca.predict}} as csv file.
}
\details{
Age group parameters are written as comma-separated file with quoted strings as row/column names.
   Each row correspond to an age group, and columns to means and standard deviations.
}
\keyword{internal}
