% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_eca.r
\name{saveCatchAtLength}
\alias{saveCatchAtLength}
\title{Save catch at Length}
\usage{
saveCatchAtLength(
  pred,
  filename,
  var = "Abundance",
  unit = "millions",
  main = ""
)
}
\arguments{
\item{pred}{as returned by \code{\link[Reca]{eca.predict}}}

\item{filename}{name of file to save to.}

\item{var}{Variable to extract. Allows for Abundance, Count or Weight}

\item{unit}{Unit for extracted variable. See \code{\link{getPlottingUnit}}}

\item{main}{Title for the analysis, to be included as comment in saved file (e.g. species and year)}
}
\description{
Write catch at length predicted by \code{\link[Reca]{eca.predict}} as csv file.
}
\details{
Catch at age length is written as comma-separated file with quoted strings as row/column names.
   Each row correspond to an age group, and columns to means and standard deviations.
   Units are controlled by parameters, and written as metainformation in a preamble identified by the comment charater '#', along with any text provided in other arguments (parameter main).
}
\keyword{internal}
