% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_eca.r
\name{saveCatchMatrix}
\alias{saveCatchMatrix}
\title{Save catch at age matrix}
\usage{
saveCatchMatrix(pred, filename, var = "Abundance", unit = "ones",
  main = "", savemeans = F)
}
\arguments{
\item{pred}{as returned by \code{\link[eca]{eca.predict}}}

\item{filename}{name of file to save to.}

\item{var}{Variable to extract. Allows for Abundance, Count or Weight}

\item{unit}{Unit for extracted variable. See \code{\link{getPlottingUnit}}}

\item{main}{Title for the analysis, to be included as comment in saved file (e.g. species and year)}

\item{savemeans}{If True, only means for each age group will be saved, otherwise extracted variable is saved for each iteration of the Monte Carlo simulation.}
}
\description{
Write catch at age predicted by \code{\link[eca]{eca.predict}} as csv file.
}
\details{
Catch at age matrix is written as comma-separated file with quoted strings as row/column names.
   Each row correspond to an age group, and columns to either means or an iteration of the Monte Carlo simulation.
   Units are controlled by parameters, and written as metainformation in a preamble identified by the comment charater '#', along with any text provided in other arguments (parameter main).
}
\keyword{internal}
