% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_eca.r
\name{saveDecomposedCatchAtLength}
\alias{saveDecomposedCatchAtLength}
\title{Make decomposed catch at length}
\usage{
saveDecomposedCatchAtLength(
  projectName,
  filename = NULL,
  decomposition = c("områdegrupperingbokmål"),
  addQuarterToDecomp = T,
  customMainAreaGrouping = NULL,
  customLocationGrouping = NULL,
  var = "Abundance",
  unit = "millions",
  main = "",
  stock = NULL
)
}
\arguments{
\item{projectName}{Name identifying StoX project}

\item{filename}{filename to write decomposed catch at length to. If NULL a filename in the projects R-report directory will be generated.}

\item{decomposition}{variables to use for decomposition, must be available for all rows in landings}

\item{addQuarterToDecomp}{workaround variable for adding quarter to decomp}

\item{customMainAreaGrouping}{optional, list mapping custom spatial groups to vectors of main area strings (2 character string <mainarea>, e.g. "12" or "09")}

\item{customLocationGrouping}{optional, list mapping custom spatial groups to vectors of location strings (5 character string <mainarea>-<location>, e.g.: "12-01" or "09-01")}

\item{var}{Variable to extract for calculation. Allows for Abundance, Count or Weight}

\item{unit}{Unit for extracted variable. See \code{\link{getPlottingUnit}}}

\item{main}{Title for the analysis, to be included as comment in saved file (e.g. species and year)}

\item{stock}{stock to extract results for if stock-splitting analysis was run. E.g. "coastal" or "atlantic"}
}
\value{
data frame with rows for each combination of decomposition variables and length groups, and columns with values or levels for length groups and decomposition variables, in addition to columns for the catch at length and standard deviation
}
\description{
Compiles catch at length table decomposed on given variables
}
\details{
Contain options for workaround variables, until reporting setup can be configured in stox.
   
   decomposition variables need not be the same as model covariates, but model covariates will be used for estimation.
   if the model covariates represent a finer decomposition of the sampling frame than the decomposition variables, interpretation is straightforward.
   if the model covariates represent a coarser decomposition of the sampling frame than the decomposition variables this implies an assumption of validity of parameteres outside the covariate combinations they are obtained for.
   
   Specifying either customMainAreaGrouping or customLocationGrouping, adds a column "spatialGroup" to the decomposition
   The spatial group is specified by the arguments 'customMainAreaGrouping' or 'customLocationGrouping' as some merge of either main-areas
   or combination of main-area location (Directorate of fisheries).
}
