% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_eca.r
\name{saveDecomposedCatchMatrix}
\alias{saveDecomposedCatchMatrix}
\title{Make decomposed catch matrix
Contain workaround variables, until reporting setup can be configured in stox.}
\usage{
saveDecomposedCatchMatrix(projectname, filename, decomposition,
  totallandings, addQuarterToDecomp = F, var = "Abundance",
  unit = "ones", main = "")
}
\arguments{
\item{decomposition}{variables to use for decomposition, must be available for all rows in landings}

\item{totallandings}{workaround variable total landings for decomposition (may be a superset for project landings). If null, project landings are used}

\item{addQuarterToDecomp}{workaround variable for adding quarter to decomp}

\item{var}{Variable to extract for calculation. Allows for Abundance, Count or Weight}

\item{unit}{Unit for extracted variable. See \code{\link{getPlottingUnit}}}

\item{main}{Title for the analysis, to be included as comment in saved file (e.g. species and year)}
}
\value{
data frame with rows for each combination of decomposition varirables, and columns (a1..an: values or levels for decomposition variables, an+1: total weight, an+2: the fraction covered by landings used for parameterization, an+3...am: columns for the mean and columns for sd for each age group
}
\description{
Compiles catch matrix decomposed on given variables
}
\details{
decomposition variables need not be the same as model covariates, but model covariates will be used for estimation.
   if the model covariates represent a finer decomposition of the sampling frame than the decomposition variables, interpretation is straightforward.
   if the model covariates represent a coarser decomposition of the sampling frame than the decomposition variables this implies an assumption of validity of parameteres outside the covariate combinations they are obtained for.
   
   if the project landings differ from totallandings, the fraction in landings for each combination of decomposition variables will be reflect that ratio, and reported estimates will reflect that of project landings
   if there are additional values or levels for the model covariates in the total landings than what exists in the project landings, NAs will be reported for means and sds.
}
\keyword{internal}
