% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_NMD.R
\name{searchNMDCruise}
\alias{searchNMDCruise}
\title{Search for a cruise given the cruise number and ship name.}
\usage{
searchNMDCruise(
  cruisenr,
  shipname = NULL,
  datasource = "biotic",
  ver = getRstoxDef("ver"),
  server = "http://tomcat7.imr.no:8080/apis/nmdapi",
  return.URL = FALSE,
  snapshot = Sys.time()
)
}
\arguments{
\item{cruisenr}{A vector of cruise numbers (see the (time consuming) command cruisenr <- getNMDinfo("cruises")$cruise).}

\item{shipname}{A vector of ship names (see the (time consuming) command shipname <- getNMDinfo("platform")$Ship_name).}

\item{datasource}{The type of data requested. Currently implemented are "echosunder" and "biotic", while "landing" and "ctd" are in the pipeline. datasource=NULL (default) returns all possible data.}

\item{ver}{The version of the API. As of 2015-05 only version 1 is available. Version 2 will include the possibility to return a list of all cruises.}

\item{server}{The URL to the server hoding the NMD data.}

\item{return.URL}{Logical: If TRUE return the URL used to search for the cruise.}

\item{snapshot}{The snapshot ID.}

\item{API}{The path to the API.}
}
\description{
The NMD API enables searching for a cruise identifyer given the cruise number and ship name.
}
\keyword{internal}
