% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_java.r
\name{setAssignments}
\alias{setAssignments}
\alias{setNASC}
\title{(Internal) Set NASC data and assignments to memory}
\usage{
setAssignments(projectName, assignments)

setNASC(projectName, process = "MeanNASC", data)
}
\arguments{
\item{projectName}{The name or full path of the project, a baseline object (as returned from getBaseline() or runBaseline()), og a project object (as returned from open).}

\item{assignments}{The modified trawl assignments (StationWeight for each combination of AssignmentID and Station).}

\item{process}{The process to set NASC values to.}

\item{data}{The modified NASC data (Value for each combination of AcoCat, SampleUnit and Layer).}
}
\value{
An assignment
}
\description{
\code{setAssignments} Sets assignments to the assignment table in the process data (in memory).
\code{setMeanNASCValues} Sets NASC or meanNASC values to the baseline data (in memory).
}
