% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_java.r
\name{setAssignments_old}
\alias{setAssignments_old}
\alias{setMeanNASCValues_old}
\title{Set and remove assignments}
\usage{
setAssignments_old(ta_table, assignments)

setMeanNASCValues_old(mtrx, tbl)
}
\arguments{
\item{ta_table}{Table 'TRAWLASSIGNMENT' in ProcessData}

\item{assignments}{Modified trawl assignments}

\item{mtrx}{The mean NASC matrix}

\item{tbl}{The table dataframe}
}
\description{
\code{setAssignments_old} Sets assignments to the assignment table in the process data (in memory).
\code{setMeanNASCValues_old} Sets NASC or meanNASC values to the baseline data (in memory).
}
\keyword{internal}

