% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_java.r
\name{setJavaMemory}
\alias{setJavaMemory}
\alias{getJavaMemory}
\title{Set the size of the Java memory}
\usage{
setJavaMemory(size = 2e+09)

getJavaMemory()
}
\arguments{
\item{size}{The size of the memory (in bytes) assigned to Java for each project}
}
\description{
Set the size of the Java memory
}
