% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_base.r
\name{setPrecisionLevel}
\alias{setPrecisionLevel}
\alias{getPrecisionLevel}
\title{Functions for setting and getting the precision level of a project.}
\usage{
setPrecisionLevel(projectName, precisionLevel)

getPrecisionLevel(projectName)
}
\arguments{
\item{projectName}{Project identifyer (see \link{\code{openProject}}).}

\item{precisionLevel}{The precision level to set to the project, where 0L represents the low precision level used prior to Rstox 1.7 and Stox 2.5, and 1L represents the 4-significant digits precision used from those versions and onward.}
}
\description{
Functions for setting and getting the precision level of a project.
}
\keyword{internal}
