% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_base.r
\name{setProjectData}
\alias{getProjectData}
\alias{loadEnv}
\alias{loadProjectData}
\alias{saveProjectData}
\alias{saveRImage}
\alias{setProjectData}
\title{Set, get, save or load project data.}
\usage{
setProjectData(projectName, var, name = NULL)

getProjectData(projectName, var = NULL)

saveProjectData(projectName, var = "all", ...)

saveRImage(projectName, var = "all", ...)

loadProjectData(projectName, var = "all", ow = FALSE, ...)

loadEnv(projectName, var = "all", ow = FALSE, ...)
}
\arguments{
\item{projectName}{The name or full path of the project, a baseline object (as returned from getBaseline() or runBaseline()), og a project object (as returned from open).}

\item{var}{for \code{setProjectData}, a project data object to be assigned to the projectData environment of the project, and for \code{getProjectData}, \code{saveProjectData} and \code{loadProjectData}, a vector of project data names to get from, save to file from, or load from file to the projectData environment, respectively.}

\item{name}{Only used in setProjectData(). The name of the project data object (such as \code{lastParameters}) to assign \code{var} to in the projectData environment of the project (overriding the name of the object \code{var}). If a specific list element inside \code{name} should be set, specify this in '...'.}

\item{...}{Used for backwards compatibility.}

\item{ow}{Logical: if TRUE overvrite objects existing in the projectData environment of the project.}
}
\description{
\code{setProjectData} Assigns an object to the projectData environment of the project. \cr \cr
\code{getProjectData} Gets an object from the projectData environment of the project. \cr \cr
\code{saveProjectData} Saves some or all objects in the projectData environment of the project to files in the output/r/data directory. \cr \cr
\code{loadProjectData} Loads some or all objects in the output/r/data directory to the projectData environment of the project. \cr \cr
\code{saveRImage} (Old function, kept for backwards compatibility) Saves the contents of the projectData environment of the project (RestoEnv[[projectName]]$projectData). \cr \cr
\code{loadEnv} (Old function, kept for backwards compatibility) Loads previously saved data to the projectData environment of the project (RestoEnv[[projectName]]$projectData). \cr \cr
}

