% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_eca.r
\name{splitBioticCC}
\alias{splitBioticCC}
\title{Splits biotic frame by otolith type}
\usage{
splitBioticCC(biotic)
}
\arguments{
\item{biotic}{frame}
}
\value{
list with members 'coastal' and 'atlantic' containing biotic entries for coastal and atlantic cod respectively
}
\description{
Splits biotic frame by otolith type
}
\keyword{internal}
