% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_eca.r
\name{splitPredCC}
\alias{splitPredCC}
\title{Splits a prediction object from running coastal cod analysis into two prediction objects, one for each stock}
\usage{
splitPredCC(pred)
}
\arguments{
\item{prediction}{object for CC analysis}
}
\value{
list with members 'coastal' and 'atlantic' containing prediction objects for coastal and atlantic cod respectively
}
\description{
Splits a prediction object from running coastal cod analysis into two prediction objects, one for each stock
}
\keyword{internal}
