% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_framework.r
\name{StationLengthDist}
\alias{StationLengthDist}
\alias{getStationLengthDist}
\title{Calculate length distributions of each combination of station and SpecCat}
\usage{
StationLengthDist(baselineOutput, BioticData = "FilterBiotic",
  LengthDistType = "PercentLengthDist", allowMissingWeight = TRUE,
  use.set = FALSE)

getStationLengthDist(BioticDataList,
  LengthDistType = "PercentLengthDist", allowMissingWeight = TRUE,
  use.set = FALSE)
}
\arguments{
\item{baselineOutput}{A list as returned from \code{\link{getBaseline}} containing biotic data with three data tables representing the levels FishStation, CatchSample and Individual.}

\item{BioticData}{The function from which the biotic data are retrieved, normally one of "ReadBioticXML" and "FilterBiotic".}

\item{LengthDistType}{The type of length distribution to use, one of "LengthDist", "NormLengthDist" and "PercentLengthDist" (see 'Details').}

\item{allowMissingWeight}{Logical: If TRUE and \code{LengthDistType} == "PercentLengthDist" accept stations with missing pairs of lengthsampleweight and (total) weight or lengthsamplecatch and (total) catch.}

\item{use.set}{For development. Remove this later!!!!!!!!!!!!!!!!!!!!!!!!!!!}

\item{BioticDataList}{A list of biotic data representing the levels FishStation, CatchSample and Individual as returned from e.g. getBaseline("Test_Rstox")$outputData$ReadBioticXML.}
}
\value{
A data table of length distributions stacked for each combination of cruise/station and SepcCat.
}
\description{
Merges two data tables with all=TRUE, while keeping only columns of the data tables with names intersecting \code{var}, and using the intersect of \code{keys} and the names of the data tables as the 'by' argument.
}
\details{
ATLE
}
