% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_spatial.r
\name{surveyPlanner}
\alias{surveyPlanner}
\alias{plotStratum}
\title{Get polygon area and convert to or from geographic and Cartesian coordinates.}
\usage{
surveyPlanner(projectName, type = c("p", "ze", "z"), bearing = "N",
  hours = 240, knots = 10, nmi = NULL, speed = NULL, seed = 0,
  dt = 1/60)

plotStratum(stratum, transect = NULL, centroid = NULL,
  transport_alpha = 0.1)
}
\arguments{
\item{x}{Either a two column matrix of x and y coordinates, indicating only one polygon, or a list of such matrices, indicating several polygons. If a list of lists of two column matrices are given, the first matrix of each list is the polygon, and the following are subtractions. \code{wkt} cal also be a wkt string such as "MULTIPOLYGON(((4 55, 15 56, 15 59, 4 59, 4 55)))".}

\item{multipolygon}{A multipolygon string.}

\item{drop}{Logical: If TRUE drop the list if only one multipolygon or only one polygon is given.}

\item{data.frame.out}{Logical: If TRUE convert the matrices to data frames with columns x and y.}
}
\description{
Get polygon area and convert to or from geographic and Cartesian coordinates.
}
