% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_eca.r
\name{temporal_workaround}
\alias{temporal_workaround}
\title{Compensates bug in stox where the covariate temporal is not set based on prosessdata when useProsessdata is chosen}
\usage{
temporal_workaround(data, processdata, sourcetype)
}
\description{
Compensates bug in stox where the covariate temporal is not set based on prosessdata when useProsessdata is chosen
}
\keyword{internal}
