% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_boostrap.r
\name{varianceEstimation}
\alias{varianceEstimation}
\title{Parametric variance estimation}
\usage{
varianceEstimation(projectName, proc = "SweptAreaDensity",
  ignore.case = TRUE)
}
\arguments{
\item{projectName}{The name or full path of the project, a baseline object (as returned from getBaseline() or runBaseline()), og a project object (as returned from open).}

\item{proc}{The process which returns the acoustic density on which the variance should be estimated.}

\item{ignore.case}{Logical: If TRUE ignore case in species category SpecCat.}
}
\value{
A list with names "Stratum" and "Total" containing dataframes for each species with the following columns: Stratum, Mean, SD (standard deviation), Variance, CV (coefficient of variation), and lower and upper 90% confidence bounds.
}
\description{
Calculates mean, variance, coefficient of variation and 90 % confidence bounds based on Jolly, G. M., & Hampton, I. (1990). A stratified random transect design for acoustic surveys of fish stocks. Canadian Journal of Fisheries and Aquatic Sciences, 47(7), 1282-1291.
}
