% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_eca.r
\name{writeRecaConfiguration}
\alias{writeRecaConfiguration}
\title{Writes RECA configuration}
\usage{
writeRecaConfiguration(GlobalParameters, Landings, WeightLength, AgeLength,
  fileobj, main = NULL)
}
\arguments{
\item{GlobalParameters}{defined in \code{\link[eca]{eca.estimate}} and \code{\link[eca]{eca.predict}}}

\item{Landings}{defined in \code{\link[eca]{eca.estimate}} and \code{\link[eca]{eca.predict}}}

\item{WeightLength}{defined in \code{\link[eca]{eca.estimate}} and \code{\link[eca]{eca.predict}}}

\item{AgeLength}{defined in \code{\link[eca]{eca.estimate}} and \code{\link[eca]{eca.predict}}}

\item{fileobj}{filename or open connection}

\item{main}{header to write before configuration}
}
\description{
Writes details about the model configuration to a text file or open connection
}
\details{
Configuration are saved reflecting the parameters as passed to \code{\link[eca]{eca.estimate}} and \code{\link[eca]{eca.predict}}, even if it is possible to have for instance the GlobalParameters differ between the two for a valid execution.
}
\keyword{internal}
