% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_spatial.r
\name{writeTransectsMaxSea}
\alias{writeTransectsMaxSea}
\alias{writeTransectsINFO}
\alias{writeTransectsTRACK}
\alias{writeTransectsGPX}
\alias{writeTransects}
\alias{getTransectFileName}
\title{Write transects generated by \code{\link{surveyPlanner}} to various report files.}
\usage{
writeTransectsMaxSea(x, projectName, dir = NULL, item.type = 257,
  item.id = 20, item.col = 1, north = "N", east = "E", digits = 3,
  byStratum = TRUE, ...)

writeTransectsINFO(x, projectName, dir = NULL, digits = 2,
  byStratum = TRUE)

writeTransectsTRACK(x, projectName, dir = NULL, digits = 5,
  byStratum = TRUE, ...)

writeTransectsGPX(x, projectName, dir = NULL, digits = 5,
  byStratum = TRUE)

writeTransects(x, projectName, dir = NULL, digits = 5, byStratum = TRUE,
  cols = NULL, keepTransport = TRUE, text = "", ext = "txt", ...)

getTransectFileName(x, projectName, text = "", ext = "txt", dir = NULL,
  byStratum = TRUE)
}
\arguments{
\item{x}{The output from \code{\link{surveyPlanner}}.}

\item{projectName}{The name or full path of the project, a baseline object (as returned from \code{\link{getBaseline}} or \code{\link{runBaseline}}, og a project object (as returned from \code{\link{openProject}}).}

\item{dir}{The path to the directory in which to put the files. If given, it takes precidence over the report directory of the project given by \code{projectName}.}

\item{item.type, item.id, item.col}{The type (default 257), id (default 20) and color (default 1) to pass to MaxSea.}

\item{north, east}{Strings denoting North and East to pass to MaxSea.}

\item{digits}{Precision to use in the reports.}

\item{byStratum}{Logical: If TRUE, create one file per stratum.}

\item{...}{Parameters passed to \code{\link[utils]{write.table}}, \code{\link[utils]{write.csv}} and \code{\link[data.table]{fwrite}}. For \code{writeTransects} thie \code{...} contains the important variables \code{sep} specifying the separator used in CSV files, and \code{dateTimeAs} which defines how to save the time in \code{\link[data.table]{fwrite}}.}

\item{cols}{The columns to write to file for \code{writeTransects}.}

\item{keepTransport}{Logical: If TRUE, save also the transport stretches to the file.}

\item{text}{A text to add to the files before the file extension.}

\item{ext}{The file extension. For \code{writeTransects} \code{ext} = "nc" implies writing NetCDF files, and otherwise a CSV files are written.}
}
\value{
The file names.
}
\description{
\code{writeTransectsToMAxSea} writes the transects of all strata to separate .asc files for import in MaxSea. \cr \cr
\code{writeTransectsINFO} writes the transects of all strata to separate hunam readable .txt files. \cr \cr
\code{writeTransectsCSV} writes the transects of all strata to separate csv files files. \cr \cr
\code{writeTransects} writes the data to CSV or NetCDF files. Only a selection of the columns are saved by default (override this by \code{cols}): The transect secifications ("stratum", "transect", "segment", "transport", "retour"), the start and stop position, time and distance ("lon_start", "lon_stop", "lat_start", "lat_stop", "time_start", "time_stop", "dist_start", "dist_stop", "segmentLengths"). By default only transects and not transport stretches are saved (override this by \code{keepTransport}). \cr \cr
}
