% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_acoustic.r
\name{wtd.strata.est}
\alias{wtd.strata.est}
\title{Weighted mean NASC by stratum}
\usage{
wtd.strata.est(tr.value, tr.dist)
}
\arguments{
\item{tr.value}{Transect NASC values.}

\item{tr.dist}{Transect distance values.}
}
\value{
list with mean and variance by stratum
}
\description{
Calculates mean by stratum (Jolly & Hampton 1990 Eq(2)) and variance by stratum (Eq3b).
}
\keyword{internal}
