% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logbooks.R
\name{adjustRecaSpatialTemporal}
\alias{adjustRecaSpatialTemporal}
\title{Adjust landings with logbooks}
\usage{
adjustRecaSpatialTemporal(
  landingsStox,
  logbook,
  processDataGear,
  processDataTemporal,
  processDataSpatial,
  gearSelection,
  temporalCovariate = "temporal",
  spatialCovariate = "spatial",
  gearCovariate = "gearfactor",
  minVesselSize = 15,
  logbookActivityCodes = c("I fiske")
)
}
\arguments{
\item{logbook}{logbooks as parser by \code{\link[Rstox]{readErsFile}}}

\item{processDataGear}{processdata gerfactor as exported from Stox-baseline}

\item{processDataTemporal}{processdata temporal}

\item{processDataSpatial}{processdata stratumpolygon as exported from Stox-baseline}

\item{gearSelection}{identifies the gear to clean for. One of the options on processDataGear$Covariate}

\item{temporalCovariate}{the name of the temporal covariate}

\item{spatialCovariate}{the name of the spatial covariate}

\item{gearCovariate}{the name of the gear covariate}

\item{minVesselSize}{minimal size in meters for vessels to include in logbook adjusted data}

\item{logbookActivityCodes}{vector specifying the activity codes to include from logbooks (field AKTIVITET).}

\item{landingStox}{landings as passed between the Reca-scripts. E.g. as saved by \code{\link[Rstox]{prepareReca}}}
}
\value{
landings, formatted as landingStox
}
\description{
Adjust landings with logbooks
}
\details{
NOTE: <note on prototyping>
 ss
}
