% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_boostrap.r
\name{bootstrapOneIteration}
\alias{bootstrapOneIteration}
\title{(Internal) Run one bootstrap iteration of biotic stations and acoustic data}
\usage{
bootstrapOneIteration(i, projectName, assignments, strata, psuNASC = NULL,
  stratumNASC = NULL, resampledNASC = NULL,
  startProcess = "TotalLengthDist", endProcess = "SuperIndAbundance",
  outProcess = "SuperIndAbundance", seedV = NULL)
}
\arguments{
\item{i}{The boostrap iteration number.}

\item{projectName}{The name or full path of the project, a baseline object (as returned from getBaseline() or runBaseline()), og a project object (as returned from open).}

\item{assignments}{Trawl assignment from baseline.}

\item{strata}{The strata of the survey.}

\item{psuNASC}{MeanNASC from baseline.}

\item{stratumNASC}{Strata NASC estimates from getNASCDistr(baseline).}

\item{resampledNASC}{Resampled NASC distribution.}

\item{seedV}{A vector of seeds. seedV[i] is used.}
}
\value{
list with (1) the abundance by length in the bootstrap run, (2) the abundance by super individuals in the bootstrap run
}
\description{
This function is used in bootstrapParallel().
}

