% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_base.r
\name{data.frame2parString}
\alias{data.frame2parString}
\alias{parString2data.frame}
\title{Get the Rstox version and the version of the Java library used by Rstox, on which StoX is built.}
\usage{
data.frame2parString(x)

parString2data.frame(x)
}
\arguments{
\item{x}{A data.frame with the parameters as columns, such as data.frame(SpecCat=c("Torsk", "Sild", ""), Alpha=runif(3), Beta=runif(3), LMin=runif(3), LMax=runif(3)).}
}
\description{
Get the Rstox version and the version of the Java library used by Rstox, on which StoX is built.
}
\examples{
df1 <- data.frame(SpecCat=c("Torsk", "Hyse", ""), Alpha=c(5,9,2), beta=c(2,2,3), LMin=c(1,2,3), LMax=c(5,7,9))
string <- data.frame2parString(df1)
df2 <- parString2data.frame(string)
df1
df2
identical(df1, df2)

}
\keyword{internal}
