% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RECA_plots.R
\name{diagnostics_model_configuration}
\alias{diagnostics_model_configuration}
\title{Plots diagnostics for model configuration. Whether all combinations of fixed effects are sampled}
\usage{
diagnostics_model_configuration(stoxexport, okcol = "green",
  wrongcol = "white", undersampledcol = "red", oktext = "OK",
  undersampledtext = "no samples", wrongtext = "no landings")
}
\description{
Plots diagnostics for model configuration. Whether all combinations of fixed effects are sampled
}
\keyword{internal}
