% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_eca.r
\name{diagnosticsRECA}
\alias{diagnosticsRECA}
\title{Generate plots for diagnosis of RECA model configuration. Compares sampling effort to fisheries along covariates selected in the model , and along some standard covariate choices if available (gear, temporal and spatial)}
\usage{
diagnosticsRECA(projectName, verbose = T, format = "png", ...)
}
\arguments{
\item{projectName}{name of stox project}

\item{verbose}{logical, if TRUE info is written to stderr()}

\item{format}{function defining filtetype for plots, supports grDevices::pdf, grDevices::png, grDevices::jpeg, grDevices::tiff, grDevices::bmp}

\item{...}{parameters passed on plot function and format}
}
\description{
Generate plots for diagnosis of RECA model configuration. Compares sampling effort to fisheries along covariates selected in the model , and along some standard covariate choices if available (gear, temporal and spatial)
}
