% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_base.r
\name{downloadProjectZip_old}
\alias{downloadProjectZip_old}
\alias{downloadProjectZip}
\title{Download a zipped StoX project to a specified project path.}
\usage{
downloadProjectZip_old(URL, projectName = NULL, projectRoot = NULL,
  cleanup = TRUE, ow = TRUE, msg = TRUE)

downloadProjectZip(URL, projectName = NULL, projectRoot = NULL,
  cleanup = TRUE, ow = TRUE, msg = TRUE)
}
\arguments{
\item{URL}{The URL of the zipped project.}

\item{cleanup}{Logical: if FALSE, the downloaded zip file is not deleted.}

\item{projectPath}{The local path to the project to be downloaded.}
}
\description{
Download a zipped StoX project to a specified project path.
}
\keyword{internal}
