% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_base.r
\name{getBioticAssignments}
\alias{getBioticAssignments}
\title{Get joined table of trawl assignments, psu and stratum}
\usage{
getBioticAssignments(projectName)
}
\arguments{
\item{projectName}{The name or full path of the project, a baseline object (as returned from getBaseline() or runBaseline()), og a project object (as returned from open).}
}
\value{
Dataframe with trawl assignments merged with psu and stratum
}
\description{
Get trawl assignments from baseline in StoX Java memory.
}
\examples{
baseline <- openProject("Test_Rstox", out="baseline")
assignments <- getBioticAssignments(baseline)

}
