% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_eca.r
\name{getLandings}
\alias{getLandings}
\title{Function for extracting the Landings object}
\usage{
getLandings(eca, ecaParameters)
}
\description{
Function for extracting the Landings object
}
\keyword{internal}
