% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_eca.r
\name{getMidSeason}
\alias{getMidSeason}
\title{Function for setting getting appropriate value for the midSeason column in Landings}
\usage{
getMidSeason(x, tz = "UTC", format = "\%d/\%m/\%Y")
}
\description{
Function for setting getting appropriate value for the midSeason column in Landings
}
\keyword{internal}
