% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_reports.r
\name{getPlottingUnit}
\alias{getPlottingUnit}
\title{Interprets a unit key stings as a scaleing factor}
\usage{
getPlottingUnit(unit = NULL, var = "Abundance", baseunit = NULL,
  implemented = c(1, 2), def.out = TRUE)
}
\arguments{
\item{unit}{A unit key string indicating the unit (see getPlottingUnit()$definitions$unlist.units for available key strings), or alternatively a numeric value giving the scaling factor.}

\item{var}{A key string indicating the variable to plot (see getPlottingUnit()$defaults$Rstox_var for available values).}

\item{baseunit}{The unit used in the data.}

\item{implemented}{An integer vector giving the inplemented variables, which are the first two ("Abundance", "Weight") in the current version of Rstox.}

\item{def.out}{Logical: if TRUE return also the defaults and definitions.}
}
\value{
a list of the following four elements: 1. the scaling factor, 2. the unit string, 3. a matrix of the default values, and 4. a matrix of the defintions.
}
\description{
Given a key string such as "milliseconds" (possibly abbreviated) or the abbreviation "ms" (identical matching), the unit (here milliseconds) and scaling factor (here 1e-3) is returned.
}
\examples{
getPlottingUnit(unit="milli", var="abund", baseunit="stox", def.out=FALSE)
getPlottingUnit(unit="milli", var="weight", baseunit="stox", def.out=FALSE)
getPlottingUnit(unit="hecto", var="weight", baseunit="stox")

}

