% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_base.r
\name{getProcess}
\alias{getProcess}
\title{Get process indices.}
\usage{
getProcess(projectName, proc = "all", modelType = "baseline",
  dropNA = TRUE, msg = TRUE, proc.out = FALSE)
}
\arguments{
\item{projectName}{The name or full path of the project, a baseline object (as returned from \code{\link{getBaseline}} or \code{\link{runBaseline}}, og a project object (as returned from \code{\link{openProject}}).}

\item{proc}{A string vector naming processes/function to find.}

\item{modelType}{The type of model to run, currently one of "baseline" and "baseline-report". First element used.}
}
\value{
Index number of the process.
}
\description{
Gets the indices of processes in the baseline model, where both process name and function name are accepted. The process WriteProcessData is ignored.
}
\examples{
openProject("Test_Rstox")
getProcess("Test_Rstox", c("SuperIndAbundance", "StratumArea"))

}
\keyword{internal}
