% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_base.r
\name{getRstoxEnv}
\alias{getRstoxEnv}
\alias{getRstoxDef}
\alias{initiateRstoxEnv}
\alias{getModelType}
\title{Set up the RstoxEnv}
\usage{
getRstoxEnv()

getRstoxDef(name)

initiateRstoxEnv()

getModelType(modelType)
}
\arguments{
\item{name}{The name of the Rstox definition to retrieve.}
}
\description{
\code{getRstoxEnv} initiates the RstoxEnv environment if not existing.
\code{getRstoxDef} gets an Rstox definition.
}
\keyword{internal}
