% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_base.r
\name{insertStratumpolygon}
\alias{insertStratumpolygon}
\alias{insertEdsupsu}
\alias{insertPsustratum}
\alias{insertSuassignment}
\alias{insertBioticassignment}
\alias{insertToProjectXML}
\alias{getTransectID}
\alias{getLogStoXid}
\title{Modify a StoX project.xml file directly.}
\usage{
insertStratumpolygon(Stratum, file)

insertEdsupsu(Log, file)

insertPsustratum(Log, file)

insertSuassignment(Log, estlayer = 1, file)

insertBioticassignment(Stations, file)

insertToProjectXML(insert, type, file)

getTransectID(Log)

getLogStoXid(Log, timevar = "start_time")
}
\arguments{
\item{Stratum}{A data.frame with column names "longitude" and "latitude" giving the stratum polygon, or a list of these, named by the stratum names.}

\item{file}{The path to the project.xml file.}

\item{Log}{A data.frame with one row per log distance, holding the following columns: "transect", "stratum", "logStoXid".}

\item{estlayer}{A vector of the estimation layers.}

\item{Stations}{A data.frame of one row per biotic station, holding the following columns: "stratum", "stationID".}

\item{insert}{A string to insert to the project.xml file.}

\item{type}{The type of data to insert to the project.xml file.}
}
\description{
\code{insertStratumpolygon} inserts stratumpolygon to the project.xml file. \cr \cr
\code{insertEdsupsu} inserts edsupsu to the project.xml file. \cr \cr
\code{insertPsustratum} inserts psustratum to the project.xml file. \cr \cr
\code{insertSuassignment} inserts suassignment to the project.xml file. \cr \cr
\code{insertBioticassignment} inserts bioticassignment to the project.xml file. \cr \cr
\code{insertToProjectXML} inserts a string to the project.xml file. \cr \cr
\code{getTransectID} returns transect IDs. \cr \cr
}
\keyword{internal}
