% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RECA_plots.R
\name{plot_cell_coverage}
\alias{plot_cell_coverage}
\title{Plot aggregated landings for different kind of sampling level in cells}
\usage{
plot_cell_coverage(eca, xlab = "sample clusteredness",
  ylab = "Fraction landed (vekt-\%)",
  titletext = "Coverage w age\\ncells (gear/temp/spatial)",
  colgood = "green4", colok = "green2", colbarely = "yellow",
  colbad = "orange", colempty = "gray", colwrong = "white",
  gooddesc = "> 1 vessel", okdesc = "> 1 catch", barelydesc = "> 0 catch",
  baddesc = "0 samples")
}
\description{
Plot aggregated landings for different kind of sampling level in cells
}
\keyword{internal}
