% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RECA_plots.R
\name{plot_gear_temporal_area}
\alias{plot_gear_temporal_area}
\title{show samples wrp common covariates gear, area and temporal}
\usage{
plot_gear_temporal_area(eca,
  titletext = "gear/temporal - area\\nlanded (kt)\\nage samples: #vessels,#catches,#individuals",
  colgood = "green4", colok = "green2", colbarely = "yellow",
  colbad = "orange", colempty = "gray", colwrong = "white")
}
\description{
show samples wrp common covariates gear, area and temporal
}
\keyword{internal}
