% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RECA_plots.R
\name{plot_RECA_results_panel}
\alias{plot_RECA_results_panel}
\title{Plots a panel of RECA results}
\usage{
plot_RECA_results_panel(pred, biotic, ...)
}
\arguments{
\item{pred}{RECA prediction object as returned by eca::eca.predict}

\item{biotic}{indiviuals as exported by stox}
}
\description{
Plots a panel of RECA results
}
\keyword{internal}
