% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_reports.r
\name{plotNASCDistribution}
\alias{plotNASCDistribution}
\title{Plot NASC distribution to file}
\usage{
plotNASCDistribution(projectName, format = "png", ...)
}
\arguments{
\item{projectName}{The name or full path of the project, a baseline object (as returned from getBaseline() or runBaseline()), og a project object (as returned from open).}

\item{format}{The file format of the saved plot (one of "png" and "tiff").}

\item{...}{Parameters passed on from other functions.}
}
\value{
Plot saved to file
}
\description{
Plots both original and resampled NASC distributions;
histogram of NASC transect means together with distribution of resampled NASC values (line).
Probability densities, component density, are plotted (so that the histogram has a total area of one).
Plot is exported to a tif- or png-file
}
\examples{
# Create the test project:
createProject("Test_Rstox", files=system.file("extdata", "Test_Rstox", package="Rstox"), ow=FALSE)
projectName <- "Test_Rstox"
# Run bootstrap before plotting:
boot <- runBootstrap(projectName, nboot=10, seed=1, 
    acousticMethod=PSU~Stratum, bioticMethod=PSU~Stratum)
plotNASCDistribution(projectName)

}
