% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_reports.r
\name{plotNASCDistribution}
\alias{plotNASCDistribution}
\title{Plot NASC distribution to file}
\usage{
plotNASCDistribution(projectName, format = "png", ...)
}
\arguments{
\item{projectName}{The name or full path of the project, a baseline object (as returned from getBaseline() or runBaseline()), og a project object (as returned from open).}

\item{format}{The file format of the saved plot, given as a string naming the function to use for saving the plot (such as bmp, jpeg, png, tiff), with \code{filename} as its first argument. Arguments fo the functions are given as \code{...}. Dimensions are defaulted to width=5000, height=3000, resolution to 500 dpi. If \code{format} has length 0, the plot is shown in the graphics window, and not saved to file.}

\item{...}{Parameters passed on from other functions.}
}
\value{
Plot saved to file
}
\description{
Plots both original and resampled NASC distributions;
histogram of NASC transect means together with distribution of resampled NASC values (line).
Probability densities, component density, are plotted (so that the histogram has a total area of one).
Plot is exported to a tif- or png-file
}
\examples{
projectName <- "Test_Rstox"
# Run bootstrap before plotting:
boot <- runBootstrap(projectName, nboot=10, seed=1, 
    acousticMethod=PSU~Stratum, bioticMethod=PSU~Stratum)
plotNASCDistribution(projectName)

}
