% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_eca.r
\name{plotRECA}
\alias{plotRECA}
\title{Defines which plots to plot to R report for RECA. Fails silently on errors.}
\usage{
plotRECA(projectName)
}
\arguments{
\item{project}{name}
}
\description{
Defines which plots to plot to R report for RECA. Fails silently on errors.
}
