% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_spatial.r
\name{polyArea}
\alias{polyArea}
\alias{wkt2MULTIPOLYGON}
\title{Polygon area}
\usage{
polyArea(wkt)

wkt2MULTIPOLYGON(wkt)
}
\arguments{
\item{wkt}{Either one two column matrix of x and y coordinates, indicating only one polygon, or a list of such matrices, indicating several polygons. If a list of lists of two column matrices are given, tne first matrix of each list is the polygon, and the following are subtractions. \code{wkt} cal also be a wkt string such as "MULTIPOLYGON(((4 55, 15 56, 15 59, 4 59, 4 55)))".}
}
\value{
\code{polyArea} returns area in nmi squared, and \code{wkt2MULTIPOLYGON} returns a MULTIPOLYGON wkt.
}
\description{
Calculate area of wkt polygon with az.eq.are projection / adapted origo. The helper function \code{wkt2MULTIPOLYGON} converts a two column matrix or a list of two column matrices to a wkt string.
}
\examples{
library(rgeos)
p1 <- cbind(c(4, 15, 15, 4, 4), c(55, 56, 59, 59, 55))
s11 <- cbind(c(7, 12, 12, 7, 7), c(56, 57, 58, 58, 56))
s12 <- cbind(c(7, 14, 5, 7), c(56, 57, 58, 56))
p2 <- cbind(20 + c(4, 15, 15, 4, 4), c(55, 56, 59, 59, 55))
s2 <- cbind(20 + c(7, 12, 12, 7, 7), c(56, 56, 58, 58, 56))
wkt <- list(
	list(p1, s11, s12),
	list(p2, s2))
wkt2MULTIPOLYGON(wkt)
plot(readWKT(wkt2MULTIPOLYGON(wkt)), col='black', pbg='white')
polyArea(wkt)

}
