% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_boostrap.r
\name{runBootstrap}
\alias{runBootstrap}
\title{Run a bootstrap in StoX}
\usage{
runBootstrap(projectName, acousticMethod = PSU ~ Stratum, bioticMethod = PSU
  ~ Stratum, nboot = 5, startProcess = "TotalLengthDist",
  endProcess = "SuperIndAbundance", seed = 1, cores = 1, msg = TRUE,
  ...)
}
\arguments{
\item{projectName}{The name or full path of the project, a baseline object (as returned from getBaseline() or runBaseline()), og a project object (as returned from open).}

\item{acousticMethod, bioticMethod}{Specification of the method to use for bootstrapping the acoustic and biotic data. Currently only one method is available for acoustic and one for biotic data: acousticMethod = PSU~Stratum, bioticMethod = PSU~Stratum. Other methods are planned in later versions, involving the levels of the data given in the below table.
\tabular{rrr}{
  Level \tab Acoustic \tab Biotic\cr
  Survey \tab Survey \tab Survey\cr
  Stratum \tab Stratum \tab Stratum\cr
  Assignment \tab Not relevant \tab Assignment of biotic station groups to acoustic PSUs\cr
  PSU \tab Acoustic data averaged over e.g. one tansect \tab Biotic station group \cr
  EDSU \tab Acoustic data averaged over e.g. one nmi \tab Biotic station\cr
  Sample \tab Ping \tab Individal catch sample
}}

\item{nboot}{Number of bootstrap replicates.}

\item{startProcess}{The start process of the bootstrapping, being the first process before which biostations has been assigned and NASC values have been calculated.}

\item{endProcess}{The end process of the bootstrapping, being the process returning a matrix containing the following columns: "Stratum", "Abundance", "weight", and grouping variables such as "age", "SpecCat", "sex".}

\item{seed}{The seed for the random number generator (used for reproducibility).}

\item{cores}{An integer giving the number of cores to run the bootstrapping over.}

\item{msg}{Logical: if TRUE print messages from runBaseline().}

\item{...}{Used for backwards compatibility.}
}
\value{
list with (1) the abundance by length in the orginal model, (2) the abundance by length in the bootstrap run, (3) the abundance by super individuals in the orginal model, (4) the abundance by super individuals in the bootstrap run
}
\description{
Resample (bootstrap) trawl stations based on swept area data and possibly also acoustic data to estimate uncertainty in estimates. By the default method (acousticMethod=PSU~Stratum, bioticMethod=PSU~Stratum), the acoustic transect values (mean NASC along transects) and biotic stations (trawls) are resampled with replacement within each stratum for each bootstrap replicate, and the StoX project rerun and super individual abundance recalculated (or the output from a different process given by \code{endProcess}).
}
\examples{
\dontrun{
b <- runBootstrap("Test_Rstox", nboot=10, seed=1, cores=1)}

}
