% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_boostrap.r
\name{runBootstrap}
\alias{runBootstrap}
\title{Run a bootstrap in StoX}
\usage{
runBootstrap(projectName, acousticMethod = NULL, bioticMethod = NULL,
  nboot = 5, seed = 1, cores = 1, msg = TRUE, ...)
}
\arguments{
\item{projectName}{The name or full path of the project, a baseline object (as returned from getBaseline() or runBaseline()), og a project object (as returned from open).}

\item{acousticMethod, bioticMethod}{Specification of the method to use for bootstrapping the acoustic and biotic data. These can be formulas or characters which can be converted to formulas, given as 'variable to bootstrap ~ level to bootstrap within'. Multiple bootstraps can be specified, such as bioticMethod=c(EDSU~Stratum, Sample~EDSU), instructing to bootstrap the EDSUs (stations) within each stratum, and also bootstrapping the individual catch samples within each EDSU. Only certain strings can be used in the formulas, as shown in the table below. The methods can also be given as functions of at least two arguments, 'projectName' and 'process', which makes modifications to the output from getBaseline(projectName, proc=process, input=NULL) and sends the modified data back to the baseline in Java memory and runs the baseline with the modified data. Using funcitons is not yet implemented. 
\tabular{ccc}{
  Level \tab Acoustic \tab Biotic\cr
  Survey \tab Survey \tab Survey\cr
  Stratum \tab Stratum \tab Stratum\cr
  Assignment \tab Not relevant \tab Assignment of biotic station groups to acoustic PSUs\cr
  PSU \tab Acoustic data averaged over e.g. one tansect \tab Biotic station group \cr
  EDSU \tab Acoustic data averaged over e.g. one nmi \tab Biotic station\cr
  Sample \tab Ping \tab Individal catch sample
}}

\item{nboot}{Number of bootstrap replicates.}

\item{seed}{The seed for the random number generator (used for reproducibility).}

\item{cores}{An integer giving the number of cores to run the bootstrapping over.}

\item{msg}{Logical: if TRUE print messages from runBaseline().}

\item{...}{Used for backwards compatibility.}
}
\value{
list with (1) the abundance by length in the orginal model, (2) the abundance by length in the bootstrap run, (3) the abundance by super individuals in the orginal model, (4) the abundance by super individuals in the bootstrap run
}
\description{
Resample (bootstrap) trawl stations based on swept area data and possibly also acoustic data to estimate uncertainty in estimates.
}
\examples{
\dontrun{
b <- runBootstrap("Test_Rstox", nboot=10, seed=1, cores=1)}

}

