% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_eca.r
\name{runRECA}
\alias{runRECA}
\title{run RECA}
\usage{
runRECA(projectName, burnin = 100, caa.burnin = 100, nSamples = 1000,
  thin = 1, fitfile = "fit", predfile = "pred", lgamodel = "log-linear",
  CC = FALSE, CCError = FALSE, seed = NULL, age.error = FALSE,
  export_only = NULL)
}
\arguments{
\item{projectName}{name of stox project}

\item{burnin}{see specification for GlobalParameters in \code{\link[eca]{eca.estimate}}}

\item{caa.burnin}{see specification for GlobalParameters in \code{\link[eca]{eca.predict}}}

\item{nSamples}{see specification for GlobalParameters in \code{\link[eca]{eca.estimate}}}

\item{thin}{see specification for GlobalParameters in \code{\link[eca]{eca.estimate}}}

\item{fitfile}{see specification for GlobalParameters in \code{\link[eca]{eca.estimate}}}

\item{predfile}{see specification for GlobalParameters in \code{\link[eca]{eca.predict}}}

\item{lgamodel}{see specification for GlobalParameters in \code{\link[eca]{eca.estimate}}}

\item{CC}{see specification for GlobalParameters in \code{\link[eca]{eca.estimate}}}

\item{CCError}{see specification for GlobalParameters in \code{\link[eca]{eca.estimate}}}

\item{seed}{see specification for GlobalParameters in \code{\link[eca]{eca.estimate}}}

\item{age.error}{see specification for GlobalParameters in \code{\link[eca]{eca.estimate}}}

\item{export_only}{if not NULL this indicates that eca should not be run, but all parameters should be exported to the file export_only}
}
\description{
Loads data produced by \code{\link{prepareRECA}} runs parameterization and makes predictions using RECA. Saves results to project data 'runRECA'
}
\details{
Most parameters to this function are appended to the argument list produced by prepareRECA and passed as argument GlobalParameters to \code{\link[eca]{eca.estimate}} and \code{\link[eca]{eca.predict}}.
    For purposes of testing and running ECA detached from StoX, Data files accepted by code{\link[eca]{eca.estimate}} and \code{\link[eca]{eca.predict}} can be exported using the option export_only. In this case the analysis is not run, but data files and parameter files are stored at the designated location.
}
