% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_eca.r
\name{runRECA}
\alias{runRECA}
\title{run RECA fit and prediction. Save results to project data 'runRECA'
parameters not described below are defined in eca.estimate and eca.predict}
\usage{
runRECA(projectName, burnin = 100, caa.burnin = 100, nSamples = 1000,
  thin = 1, fitfile = "fit", predfile = "pred", lgamodel = "log-linear",
  CC = FALSE, CCError = FALSE, seed = NULL, age.error = FALSE,
  export_only = NULL)
}
\arguments{
\item{export_only}{if not NULL this indicates that eca should not be run, but all parameters should be exported to the file export_only}
}
\description{
run RECA fit and prediction. Save results to project data 'runRECA'
parameters not described below are defined in eca.estimate and eca.predict
}
