% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_base.r
\name{sampleSorted}
\alias{sampleSorted}
\title{Sample a vector after sorting the vector and applying the seed}
\usage{
sampleSorted(x, size, seed = 0, by = NULL, replace = TRUE,
  sorted = TRUE, drop = FALSE)
}
\arguments{
\item{x}{A vector or a single integer.}

\item{size}{A non-negative integer giving the number of items to choose.}

\item{seed}{The seed to apply before sampling.}

\item{by}{The name of the column to sample by when sampling rows of a data frame.}

\item{replace}{Should sampling be with replacement?}

\item{sorted}{Should the data be sorted prior to sampling?}

\item{drop}{Should data frames be dropped when sampling?}
}
\description{
Sample a vector after sorting the vector and applying the seed
}
\keyword{internal}
