% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_eca.r
\name{saveDecomposedCatchMatrix}
\alias{saveDecomposedCatchMatrix}
\title{Make decomposed catch matrix}
\usage{
saveDecomposedCatchMatrix(
  projectName,
  filename = NULL,
  decomposition = c("områdegrupperingbokmål"),
  addQuarterToDecomp = T,
  var = "Abundance",
  unit = "millions",
  plusgr = NULL,
  main = ""
)
}
\arguments{
\item{projectName}{Name identifying StoX project}

\item{filename}{filename to write decomposed catch matrix to. If NULL a filename in the projects R-report directory will be generated.}

\item{decomposition}{variables to use for decomposition, must be available for all rows in landings}

\item{addQuarterToDecomp}{workaround variable for adding quarter to decomp}

\item{var}{Variable to extract for calculation. Allows for Abundance, Count or Weight}

\item{unit}{Unit for extracted variable. See \code{\link{getPlottingUnit}}}

\item{plusgr}{Lower age in plusgr for tabulation. If NULL plusgr is not used.}

\item{main}{Title for the analysis, to be included as comment in saved file (e.g. species and year)}
}
\value{
data frame with rows for each combination of decomposition varirables, and columns (a1..an: values or levels for decomposition variables, an+1: total weight, an+2: the fraction covered by landings used for parameterization, an+3...am: columns for the mean and columns for sd for each age group
}
\description{
Compiles catch matrix decomposed on given variables
}
\details{
Contain options for workaround variables, until reporting setup can be configured in stox.
   
   decomposition variables need not be the same as model covariates, but model covariates will be used for estimation.
   if the model covariates represent a finer decomposition of the sampling frame than the decomposition variables, interpretation is straightforward.
   if the model covariates represent a coarser decomposition of the sampling frame than the decomposition variables this implies an assumption of validity of parameteres outside the covariate combinations they are obtained for.
}
