% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_java.r
\name{setJavaMemory}
\alias{setJavaMemory}
\alias{getJavaMemory}
\title{Get or set the size of the Java memory}
\usage{
setJavaMemory(size = 6e+09)

getJavaMemory()
}
\arguments{
\item{size}{The size of the memory (in bytes) assigned to Java for each project. This value should not be as large as the total memory of the system.}
}
\description{
Get or set the size of the Java memory
}
\details{
The function \code{setJavaMemory} sets the memory reserved for the Java virtual machine. Use \code{getJavaMemory()} to get the current value. The default is 2e9 (2 gigabytes), which should be sufficient in most cases. See \code{\link{runBootstrap}} for how to reduce the memory occupied by Java during parallel bootstrap operations.
}
