% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_boostrap.r
\name{varianceEstimation}
\alias{varianceEstimation}
\title{Parametric variance estimation}
\usage{
varianceEstimation(
  projectName,
  proc = "SweptAreaDensity",
  ignore.case = TRUE,
  na.rm = TRUE,
  out = c("bySpecCat", "byLevel", "listByLevel")
)
}
\arguments{
\item{projectName}{The name or full path of the project, a baseline object (as returned from \code{\link{getBaseline}} or \code{\link{runBaseline}}, og a project object (as returned from \code{\link{openProject}}).}

\item{proc}{The process which returns the acoustic density on which the variance should be estimated.}

\item{ignore.case}{Logical: If TRUE ignore case in species category SpecCat.}

\item{na.rm}{Logical: If FALSE return NA for all variance and mean statistics in the presence of NA in at lesat one stratum.}

\item{out}{A string specifying the format of the output (first element used). (1) If out = "bySpecCat", return a list with one element per species category SpecCat, with sublists for the total survey and for each stratum. (2) If out = "byLevel", return a list with elements 'Total' and 'Stratum', where all species categories are merged in one data frame. (3) If out = "listByLevel", return a list with elements 'Total' and 'Stratum', with sublists for all species categories.}
}
\value{
A list with names "Stratum" and "Total" containing dataframes for each species with the following columns: Stratum, Mean, SD (standard deviation), Variance, CV (coefficient of variation), and lower and upper 90% confidence bounds.
}
\description{
Calculates mean, variance, coefficient of variation and 90 % confidence bounds based on Jolly, G. M., & Hampton, I. (1990). A stratified random transect design for acoustic surveys of fish stocks. Canadian Journal of Fisheries and Aquatic Sciences, 47(7), 1282-1291.
}
\examples{
\dontrun{
# Download an example swept area project:
# pr <- createProject(
#     "ftp://ftp.imr.no/StoX/Example\%20projects/Example_sweptarea_BarentsSea_cod_1999.zip"
# )
# v <- varianceEstimation(pr)
}

}
