% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstox_spatial.r
\name{writeTransectsToMAxSea}
\alias{writeTransectsToMAxSea}
\alias{writeTransectsINFO}
\alias{writeTransectsCSV}
\alias{writeTransectsNC}
\title{Write transects generated by \code{\link{surveyPlanner}} to various report files.}
\usage{
writeTransectsToMAxSea(x, projectName, item.type = 257, item.id = 20,
  item.col = 1, north = "N", east = "E", digits = 3, ...)

writeTransectsINFO(x, projectName, digits = 2)

writeTransectsCSV(x, projectName, ...)

writeTransectsNC(x, projectName, ...)
}
\arguments{
\item{x}{The output from \code{\link{surveyPlanner}}.}

\item{projectName}{The name or full path of the project, a baseline object (as returned from \code{\link{getBaseline}} or \code{\link{runBaseline}}, og a project object (as returned from \code{\link{openProject}}).}

\item{item.type, item.id, item.col}{The type (default 257), id (default 20) and color (default 1) to pass to MaxSea.}

\item{north, east}{Strings denoting North and East.}

\item{digits}{Precision to use in the reports.}

\item{...}{Parameters passed to \code{write.table}, and \code{write.csv}.}
}
\value{
The file names.
}
\description{
\code{writeTransectsToMAxSea} writes the transects of all strata to separate .asc files for import in MaxSea. \cr \cr
\code{writeTransectsINFO} writes the transects of all strata to separate hunam readable .txt files. \cr \cr
\code{writeTransectsCSV} writes the transects of all strata to separate csv files files. \cr \cr
\code{writeTransectsNC} writes the transect mid positions and time to a NetCDF4 file. \cr \cr
}
